@@ -1,4 +1,4 @@
-// Package testing provides testing utilities for the Gollm library.
+// Package assess provides testing utilities for the Gollm library.
 package assess
 
 import (
@@ -11,10 +11,27 @@ import (
 	"testing"
 	"time"
 
-	"github.com/teilomillet/gollm"
-	"github.com/teilomillet/gollm/config"
-	"github.com/teilomillet/gollm/llm"
+	"github.com/modelcontextprotocol/go-sdk/jsonschema"
 	"golang.org/x/time/rate"
+
+	"github.com/weave-labs/gollm"
+	"github.com/weave-labs/gollm/config"
+	"github.com/weave-labs/gollm/internal/models"
+	"github.com/weave-labs/gollm/llm"
+	"github.com/weave-labs/gollm/providers"
+)
+
+// Constants for default values
+const (
+	DefaultTestTimeout     = 30 * time.Second
+	DefaultBatchTimeout    = 10 * time.Minute
+	DefaultMaxParallel     = 5
+	DefaultMaxRetries      = 3
+	DefaultRetryDelay      = 2 * time.Second
+	AverageWeight          = 2
+	PercentageMultiplier   = 100
+	DefaultAnthropicTokens = 1000
+	DefaultOpenAITokens    = 500
 )
 
 // BatchTestConfig configures batch test execution
@@ -50,21 +67,22 @@ type TestProvider struct {
 
 // TestCase represents a single test scenario
 type TestCase struct {
-	Name           string
-	Input          string
-	SystemPrompt   string
-	ExpectedSchema interface{}
-	Timeout        time.Duration
-	Validations    []ValidationFunc
-	options        map[string]interface{}
-	directives     []string
-	context        string
-	maxLength      int
-	examples       []string
-	tools          []gollm.Tool
-	toolChoice     string
-	messages       []gollm.PromptMessage
-	output         string
+	Name               string
+	Input              string
+	SystemPrompt       string
+	ExpectedSchema     *jsonschema.Schema
+	ExpectedSchemaType any // Used for JSON schema validation
+	Timeout            time.Duration
+	Validations        []ValidationFunc
+	options            map[string]any
+	directives         []string
+	context            string
+	maxLength          int
+	examples           []string
+	tools              []models.Tool
+	toolChoice         string
+	messages           []gollm.PromptMessage
+	output             string
 }
 
 // ValidationFunc is a function type for custom validations
@@ -91,7 +109,9 @@ type TestMetrics struct {
 	StartTime     time.Time
 }
 
+// NewTest creates a new test runner instance
 func NewTest(t *testing.T) *TestRunner {
+	t.Helper()
 	return &TestRunner{
 		t: t,
 		metrics: &TestMetrics{
@@ -104,6 +124,7 @@ func NewTest(t *testing.T) *TestRunner {
 	}
 }
 
+// WithProvider adds a provider to test against.
 func (tr *TestRunner) WithProvider(name, model string) *TestRunner {
 	tr.providers = append(tr.providers, TestProvider{
 		Name:    name,
@@ -113,63 +134,79 @@ func (tr *TestRunner) WithProvider(name, model string) *TestRunner {
 	return tr
 }
 
-func (tr *TestRunner) WithProviders(providers map[string]string) *TestRunner {
-	for name, model := range providers {
+// WithProviders configures the test runner with multiple providers
+func (tr *TestRunner) WithProviders(providerList map[string]string) *TestRunner {
+	for name, model := range providerList {
 		tr.WithProvider(name, model)
 	}
 	return tr
 }
 
+// WithConfig sets the configuration for the test runner
 func (tr *TestRunner) WithConfig(cfg *config.Config) *TestRunner {
 	tr.config = cfg
 	return tr
 }
 
+// WithSystemPrompt sets the system prompt for the test case
 func (tc *TestCase) WithSystemPrompt(prompt string) *TestCase {
 	tc.SystemPrompt = prompt
 	return tc
 }
 
+// WithTimeout sets the timeout duration for the test case
 func (tc *TestCase) WithTimeout(timeout time.Duration) *TestCase {
 	tc.Timeout = timeout
 	return tc
 }
 
-func (tc *TestCase) ExpectSchema(schema interface{}) *TestCase {
+// SetExpectedSchema sets the expected schema for validation
+func SetExpectedSchema[T any](tc *TestCase) *TestCase {
+	schema, err := jsonschema.For[T]()
+	if err != nil {
+		panic(fmt.Errorf("failed to get schema for type %T: %w", tc, err))
+	}
+
 	tc.ExpectedSchema = schema
+	tc.ExpectedSchemaType = *new(T)
+
 	return tc
 }
 
+// Validate adds a validation function to the test case
 func (tc *TestCase) Validate(fn ValidationFunc) *TestCase {
 	tc.Validations = append(tc.Validations, fn)
 	return tc
 }
 
-func (tc *TestCase) WithOption(key string, value interface{}) *TestCase {
+// WithOption sets a test case option
+func (tc *TestCase) WithOption(key string, value any) *TestCase {
 	if tc.options == nil {
-		tc.options = make(map[string]interface{})
+		tc.options = make(map[string]any)
 	}
 	tc.options[key] = value
 	return tc
 }
 
+// AddCase adds a new test case to the runner
 func (tr *TestRunner) AddCase(name, input string) *TestCase {
 	tc := &TestCase{
 		Name:    name,
 		Input:   input,
-		Timeout: 30 * time.Second,
-		options: make(map[string]interface{}),
+		Timeout: DefaultTestTimeout,
+		options: make(map[string]any),
 	}
 	tr.cases = append(tr.cases, tc)
 	return tc
 }
 
+// WithBatchConfig configures batch test execution settings
 func (tr *TestRunner) WithBatchConfig(cfg BatchTestConfig) *TestRunner {
 	if cfg.MaxParallel <= 0 {
-		cfg.MaxParallel = 5 // default
+		cfg.MaxParallel = DefaultMaxParallel // default
 	}
 	if cfg.BatchTimeout <= 0 {
-		cfg.BatchTimeout = 10 * time.Minute // default
+		cfg.BatchTimeout = DefaultBatchTimeout // default
 	}
 	tr.batchCfg = &cfg
 	tr.batchMetrics = &BatchMetrics{
@@ -179,6 +216,7 @@ func (tr *TestRunner) WithBatchConfig(cfg BatchTestConfig) *TestRunner {
 	return tr
 }
 
+// GetBatchMetrics returns the batch execution metrics
 func (tr *TestRunner) GetBatchMetrics() *BatchMetrics {
 	return tr.batchMetrics
 }
@@ -209,12 +247,29 @@ func (tr *TestRunner) printErrorSummary() {
 	tr.t.Log("==================")
 }
 
+// testResult represents the result of a single test execution
+type testResult struct {
+	provider string
+	testCase string
+	duration time.Duration
+	err      error
+	response string
+}
+
+// workerState tracks concurrency state across workers
+type workerState struct {
+	current int
+	max     int
+	mu      sync.Mutex
+}
+
+// RunBatch executes all test cases in parallel batches
 func (tr *TestRunner) RunBatch(ctx context.Context) {
 	if tr.batchCfg == nil {
 		tr.batchCfg = &BatchTestConfig{
 			EnableBatch:  true,
-			MaxParallel:  5,
-			BatchTimeout: 10 * time.Minute,
+			MaxParallel:  DefaultMaxParallel,
+			BatchTimeout: DefaultBatchTimeout,
 		}
 	}
 
@@ -227,89 +282,11 @@ func (tr *TestRunner) RunBatch(ctx context.Context) {
 	// Create worker pool
 	var wg sync.WaitGroup
 	semaphore := make(chan struct{}, tr.batchCfg.MaxParallel)
-	currentConcurrent := 0
-	maxConcurrent := 0
-
-	var concurrencyMu sync.Mutex
-
-	// Create channels for real-time updates
-	type testResult struct {
-		provider string
-		testCase string
-		duration time.Duration
-		err      error
-		response string
-	}
+	workerState := &workerState{}
 	results := make(chan testResult, len(tr.providers)*len(tr.cases))
 
-	for _, provider := range tr.providers {
-		client := tr.setupClient(provider)
-		tr.t.Logf("Initialized client for provider: %s", provider.Name)
-
-		for _, tc := range tr.cases {
-			wg.Add(1)
-			go func(p TestProvider, testCase *TestCase) {
-				defer wg.Done()
-
-				// Acquire semaphore
-				semaphore <- struct{}{}
-				concurrencyMu.Lock()
-				currentConcurrent++
-				if currentConcurrent > maxConcurrent {
-					maxConcurrent = currentConcurrent
-				}
-				concurrencyMu.Unlock()
-
-				tr.t.Logf("Starting test case [%s] for provider [%s]", testCase.Name, p.Name)
-
-				// Apply rate limiting if configured
-				if tr.batchCfg.RateLimit != nil {
-					if err := tr.batchCfg.RateLimit.Wait(ctx); err != nil {
-						tr.recordError(p.Name, fmt.Errorf("rate limit wait failed: %w", err))
-						tr.t.Logf("Rate limit wait error for provider %s: %v", p.Name, err)
-					}
-				}
-
-				// Run the test case
-				start := time.Now()
-				var testErr error
-				var response string
-
-				// Create a sub-test for proper test organization
-				tr.t.Run(fmt.Sprintf("%s/%s", p.Name, testCase.Name), func(t *testing.T) {
-					response, testErr = tr.runBatchCase(ctx, t, client, p, testCase)
-				})
-
-				duration := time.Since(start)
-
-				// Send result through channel
-				results <- testResult{
-					provider: p.Name,
-					testCase: testCase.Name,
-					duration: duration,
-					err:      testErr,
-					response: response,
-				}
-
-				// Update provider latency metrics
-				tr.mu.Lock()
-				if existing, ok := tr.batchMetrics.BatchTiming.ProviderLatency[p.Name]; ok {
-					tr.batchMetrics.BatchTiming.ProviderLatency[p.Name] = (existing + duration) / 2
-				} else {
-					tr.batchMetrics.BatchTiming.ProviderLatency[p.Name] = duration
-				}
-				tr.mu.Unlock()
-
-				// Release semaphore
-				concurrencyMu.Lock()
-				currentConcurrent--
-				concurrencyMu.Unlock()
-				<-semaphore
-
-				tr.t.Logf("Completed test case [%s] for provider [%s] in %v", testCase.Name, p.Name, duration)
-			}(provider, tc)
-		}
-	}
+	// Launch test workers
+	tr.launchTestWorkers(ctx, &wg, semaphore, workerState, results)
 
 	// Start a goroutine to close results channel when all tests complete
 	go func() {
@@ -317,7 +294,105 @@ func (tr *TestRunner) RunBatch(ctx context.Context) {
 		close(results)
 	}()
 
-	// Process results as they come in
+	// Process results
+	tr.processTestResults(results)
+
+	// Finalize metrics
+	tr.finalizeBatchMetrics(workerState.max)
+}
+
+// launchTestWorkers starts test execution workers
+func (tr *TestRunner) launchTestWorkers(
+	ctx context.Context,
+	wg *sync.WaitGroup,
+	semaphore chan struct{},
+	state *workerState,
+	results chan testResult,
+) {
+	for _, provider := range tr.providers {
+		client := tr.setupClient(provider)
+		tr.t.Logf("Initialized client for provider: %s", provider.Name)
+
+		for _, tc := range tr.cases {
+			wg.Add(1)
+			go tr.runTestWorker(ctx, wg, semaphore, state, results, provider, tc, client)
+		}
+	}
+}
+
+// runTestWorker executes a single test in a worker goroutine
+func (tr *TestRunner) runTestWorker(
+	ctx context.Context,
+	wg *sync.WaitGroup,
+	semaphore chan struct{},
+	state *workerState,
+	results chan testResult,
+	p TestProvider,
+	testCase *TestCase,
+	client llm.LLM,
+) {
+	defer wg.Done()
+
+	// Acquire semaphore
+	semaphore <- struct{}{}
+	state.mu.Lock()
+	state.current++
+	if state.current > state.max {
+		state.max = state.current
+	}
+	state.mu.Unlock()
+
+	tr.t.Logf("Starting test case [%s] for provider [%s]", testCase.Name, p.Name)
+
+	// Apply rate limiting if configured
+	if tr.batchCfg.RateLimit != nil {
+		if err := tr.batchCfg.RateLimit.Wait(ctx); err != nil {
+			tr.recordError(p.Name, fmt.Errorf("rate limit wait failed: %w", err))
+			tr.t.Logf("Rate limit wait error for provider %s: %v", p.Name, err)
+		}
+	}
+
+	// Run the test case
+	start := time.Now()
+	var testErr error
+	var response string
+
+	// Create a sub-test for proper test organization
+	tr.t.Run(fmt.Sprintf("%s/%s", p.Name, testCase.Name), func(t *testing.T) {
+		response, testErr = tr.runBatchCase(ctx, t, client, testCase)
+	})
+
+	duration := time.Since(start)
+
+	// Send result through channel
+	results <- testResult{
+		provider: p.Name,
+		testCase: testCase.Name,
+		duration: duration,
+		err:      testErr,
+		response: response,
+	}
+
+	// Update provider latency metrics
+	tr.mu.Lock()
+	if existing, ok := tr.batchMetrics.BatchTiming.ProviderLatency[p.Name]; ok {
+		tr.batchMetrics.BatchTiming.ProviderLatency[p.Name] = (existing + duration) / AverageWeight
+	} else {
+		tr.batchMetrics.BatchTiming.ProviderLatency[p.Name] = duration
+	}
+	tr.mu.Unlock()
+
+	// Release semaphore
+	state.mu.Lock()
+	state.current--
+	state.mu.Unlock()
+	<-semaphore
+
+	tr.t.Logf("Completed test case [%s] for provider [%s] in %v", testCase.Name, p.Name, duration)
+}
+
+// processTestResults handles incoming test results
+func (tr *TestRunner) processTestResults(results chan testResult) {
 	completedTests := 0
 	totalTests := len(tr.providers) * len(tr.cases)
 	for result := range results {
@@ -328,12 +403,21 @@ func (tr *TestRunner) RunBatch(ctx context.Context) {
 		} else {
 			tr.t.Logf("✓ [%s/%s] Completed in %v", result.provider, result.testCase, result.duration)
 		}
-		tr.t.Logf("Progress: %d/%d tests completed (%d%%)", completedTests, totalTests, (completedTests*100)/totalTests)
+		tr.t.Logf(
+			"Progress: %d/%d tests completed (%d%%)",
+			completedTests,
+			totalTests,
+			(completedTests*PercentageMultiplier)/totalTests,
+		)
 	}
+}
 
-	// Record final metrics
+// finalizeBatchMetrics records final execution metrics
+func (tr *TestRunner) finalizeBatchMetrics(maxConcurrent int) {
 	tr.batchMetrics.BatchTiming.EndTime = time.Now()
-	tr.batchMetrics.BatchTiming.TotalDuration = tr.batchMetrics.BatchTiming.EndTime.Sub(tr.batchMetrics.BatchTiming.StartTime)
+	tr.batchMetrics.BatchTiming.TotalDuration = tr.batchMetrics.BatchTiming.EndTime.Sub(
+		tr.batchMetrics.BatchTiming.StartTime,
+	)
 	tr.batchMetrics.ConcurrencyStats.MaxConcurrent = maxConcurrent
 
 	tr.t.Logf("Batch execution completed in %v", tr.batchMetrics.BatchTiming.TotalDuration)
@@ -344,11 +428,12 @@ func (tr *TestRunner) RunBatch(ctx context.Context) {
 }
 
 // Helper method to run a single batch test case
-func (tr *TestRunner) runBatchCase(ctx context.Context, t *testing.T, client llm.LLM, provider TestProvider, tc *TestCase) (string, error) {
+func (tr *TestRunner) runBatchCase(ctx context.Context, t *testing.T, client llm.LLM, tc *TestCase) (string, error) {
+	t.Helper()
 	ctx, cancel := context.WithTimeout(ctx, tc.Timeout)
 	defer cancel()
 
-	prompt := client.NewPrompt(tc.Input)
+	prompt := llm.NewPrompt(tc.Input)
 	if tc.SystemPrompt != "" {
 		prompt.SystemPrompt = tc.SystemPrompt
 	}
@@ -360,33 +445,27 @@ func (tr *TestRunner) runBatchCase(ctx context.Context, t *testing.T, client llm
 		}
 	}
 
-	var response string
+	var response *providers.Response
 	var err error
 
-	if tc.ExpectedSchema != nil {
-		// Use JSON schema validation for all providers
-		response, err = client.GenerateWithSchema(ctx, prompt, tc.ExpectedSchema)
-	} else {
-		response, err = client.Generate(ctx, prompt)
-	}
+	response, err = client.Generate(ctx, prompt)
 
 	if err != nil {
-		return "", err
+		return "", fmt.Errorf("failed to generate response: %w", err)
 	}
 
 	// Run validations
 	for _, validate := range tc.Validations {
-		if err := validate(response); err != nil {
+		if err := validate(response.AsText()); err != nil {
 			t.Errorf("Validation failed: %v", err)
 		}
 	}
 
-	return response, nil
+	return response.AsText(), nil
 }
 
 func (tr *TestRunner) setupClient(provider TestProvider) llm.LLM {
-	// Get API key from environment
-	apiKeyEnv := fmt.Sprintf("%s_API_KEY", strings.ToUpper(provider.Name))
+	apiKeyEnv := strings.ToUpper(provider.Name) + "_API_KEY"
 	apiKey := os.Getenv(apiKeyEnv)
 	if apiKey == "" {
 		tr.t.Skipf("Skipping tests for %s: %s environment variable not set", provider.Name, apiKeyEnv)
@@ -397,22 +476,22 @@ func (tr *TestRunner) setupClient(provider TestProvider) llm.LLM {
 		gollm.SetProvider(provider.Name),
 		gollm.SetModel(provider.Model),
 		gollm.SetAPIKey(apiKey),
-		gollm.SetMaxRetries(3),
-		gollm.SetRetryDelay(time.Second * 2),
+		gollm.SetMaxRetries(DefaultMaxRetries),
+		gollm.SetRetryDelay(DefaultRetryDelay),
 		gollm.SetLogLevel(gollm.LogLevelInfo),
 		gollm.SetEnableCaching(true),
-		gollm.SetTimeout(30 * time.Second),
+		gollm.SetTimeout(DefaultTestTimeout),
 	}
 
 	// Add provider-specific settings
 	switch provider.Name {
 	case "anthropic":
 		opts = append(opts,
-			gollm.SetMaxTokens(1000),
+			gollm.SetMaxTokens(DefaultAnthropicTokens),
 		)
 	case "openai":
 		opts = append(opts,
-			gollm.SetMaxTokens(500),
+			gollm.SetMaxTokens(DefaultOpenAITokens),
 		)
 	}
 
@@ -433,12 +512,9 @@ func (tr *TestRunner) setupClient(provider TestProvider) llm.LLM {
 	return client
 }
 
-func (tr *TestRunner) runCase(ctx context.Context, t *testing.T, client llm.LLM, provider TestProvider, tc *TestCase) {
-	ctx, cancel := context.WithTimeout(ctx, tc.Timeout)
-	defer cancel()
-
-	// Create prompt with all options
-	promptOpts := []gollm.PromptOption{}
+// buildPromptOptions creates prompt options from test case configuration
+func (tc *TestCase) buildPromptOptions() []gollm.PromptOption {
+	var promptOpts []gollm.PromptOption
 
 	if tc.SystemPrompt != "" {
 		promptOpts = append(promptOpts, gollm.WithSystemPrompt(tc.SystemPrompt, gollm.CacheTypeEphemeral))
@@ -480,6 +556,16 @@ func (tr *TestRunner) runCase(ctx context.Context, t *testing.T, client llm.LLM,
 		promptOpts = append(promptOpts, gollm.WithOutput(tc.output))
 	}
 
+	return promptOpts
+}
+
+func (tr *TestRunner) runCase(ctx context.Context, t *testing.T, client llm.LLM, provider TestProvider, tc *TestCase) {
+	t.Helper()
+	ctx, cancel := context.WithTimeout(ctx, tc.Timeout)
+	defer cancel()
+
+	// Create prompt with all options
+	promptOpts := tc.buildPromptOptions()
 	prompt := gollm.NewPrompt(tc.Input, promptOpts...)
 
 	// Apply test case options
@@ -490,14 +576,10 @@ func (tr *TestRunner) runCase(ctx context.Context, t *testing.T, client llm.LLM,
 	}
 
 	start := time.Now()
-	var response string
+	var response *providers.Response
 	var err error
 
-	if tc.ExpectedSchema != nil {
-		response, err = client.GenerateWithSchema(ctx, prompt, tc.ExpectedSchema)
-	} else {
-		response, err = client.Generate(ctx, prompt)
-	}
+	response, err = client.Generate(ctx, prompt)
 
 	duration := time.Since(start)
 	tr.metrics.ResponseTimes[provider.Name] = append(tr.metrics.ResponseTimes[provider.Name], duration)
@@ -510,7 +592,7 @@ func (tr *TestRunner) runCase(ctx context.Context, t *testing.T, client llm.LLM,
 
 	// Run validations
 	for _, validate := range tc.Validations {
-		if err := validate(response); err != nil {
+		if err := validate(response.AsText()); err != nil {
 			t.Errorf("Validation failed: %v", err)
 		}
 	}
@@ -529,7 +611,7 @@ func getDefaultHeaders(provider string) map[string]string {
 	return headers
 }
 
-// Common validation functions
+// ExpectContains Common validation functions
 func ExpectContains(substr string) ValidationFunc {
 	return func(response string) error {
 		if !strings.Contains(response, substr) {
@@ -539,11 +621,12 @@ func ExpectContains(substr string) ValidationFunc {
 	}
 }
 
+// ExpectMatches creates a validation function that checks if response matches a regex pattern
 func ExpectMatches(pattern string) ValidationFunc {
 	return func(response string) error {
 		matched, err := regexp.MatchString(pattern, response)
 		if err != nil {
-			return fmt.Errorf("invalid pattern %q: %v", pattern, err)
+			return fmt.Errorf("invalid pattern %q: %w", pattern, err)
 		}
 		if !matched {
 			return fmt.Errorf("expected response to match pattern %q", pattern)
@@ -552,6 +635,7 @@ func ExpectMatches(pattern string) ValidationFunc {
 	}
 }
 
+// Run executes all test cases sequentially
 func (tr *TestRunner) Run(ctx context.Context) {
 	for _, provider := range tr.providers {
 		tr.t.Run(provider.Name, func(t *testing.T) {
@@ -586,9 +670,9 @@ func (tc *TestCase) WithDirectives(directives []string) *TestCase {
 }
 
 // WithContext adds context to the test case
-func (tc *TestCase) WithContext(context string) *TestCase {
-	tc.context = context
-	tc.options["context"] = context
+func (tc *TestCase) WithContext(ctx string) *TestCase {
+	tc.context = ctx
+	tc.options["context"] = ctx
 	return tc
 }
 
@@ -607,7 +691,7 @@ func (tc *TestCase) WithExamples(examples []string) *TestCase {
 }
 
 // WithTools configures available tools for the test case
-func (tc *TestCase) WithTools(tools []gollm.Tool) *TestCase {
+func (tc *TestCase) WithTools(tools []models.Tool) *TestCase {
 	tc.tools = tools
 	tc.options["tools"] = tools
 	return tc
