@@ -11,10 +11,12 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
-	"github.com/teilomillet/gollm"
-	"github.com/teilomillet/gollm/config"
-	"github.com/teilomillet/gollm/providers"
-	"github.com/teilomillet/gollm/utils"
+	"github.com/stretchr/testify/require"
+
+	"github.com/weave-labs/gollm"
+	"github.com/weave-labs/gollm/config"
+	"github.com/weave-labs/gollm/internal/logging"
+	"github.com/weave-labs/gollm/providers"
 )
 
 // MockProvider implements the Provider interface for testing
@@ -34,12 +36,20 @@ func (m *MockProvider) Endpoint() string {
 
 func (m *MockProvider) Headers() map[string]string {
 	args := m.Called()
-	return args.Get(0).(map[string]string)
+	headers, ok := args.Get(0).(map[string]string)
+	if !ok {
+		panic("Headers() must return map[string]string")
+	}
+	return headers
 }
 
-func (m *MockProvider) PrepareRequest(prompt string, options map[string]interface{}) ([]byte, error) {
+func (m *MockProvider) PrepareRequest(prompt string, options map[string]any) ([]byte, error) {
 	args := m.Called(prompt, options)
-	return args.Get(0).([]byte), args.Error(1)
+	body, ok := args.Get(0).([]byte)
+	if !ok {
+		panic("PrepareRequest() must return []byte")
+	}
+	return body, args.Error(1)
 }
 
 func (m *MockProvider) ParseResponse(body []byte) (string, error) {
@@ -52,15 +62,15 @@ func (m *MockProvider) SupportsJSONSchema() bool {
 	return args.Bool(0)
 }
 
-func (m *MockProvider) SetOption(key string, value interface{}) {
+func (m *MockProvider) SetOption(key string, value any) {
 	m.Called(key, value)
 }
 
-func (m *MockProvider) SetDefaultOptions(config *config.Config) {
-	m.Called(config)
+func (m *MockProvider) SetDefaultOptions(cfg *config.Config) {
+	m.Called(cfg)
 }
 
-func (m *MockProvider) SetLogger(logger utils.Logger) {
+func (m *MockProvider) SetLogger(logger logging.Logger) {
 	m.Called(logger)
 }
 
@@ -78,7 +88,7 @@ func TestStructuredOutput(t *testing.T) {
 	// Create and configure mock provider
 	mockProvider := new(MockProvider)
 	mockProvider.On("Name").Return("mock")
-	mockProvider.On("SupportsJSONSchema").Return(true)
+	mockProvider.On("SupportsStructuredResponse").Return(true)
 	mockProvider.On("SetOption", mock.Anything, mock.Anything).Return()
 	mockProvider.On("SetDefaultOptions", mock.Anything).Return()
 	mockProvider.On("SetLogger", mock.Anything).Return()
@@ -101,7 +111,7 @@ func TestStructuredOutput(t *testing.T) {
 		gollm.SetAPIKey("test-key"),
 		gollm.WithProviderRegistry(registry),
 	)
-	assert.NoError(t, err)
+	require.NoError(t, err)
 
 	prompt := gollm.NewPrompt("List the top 3 benefits of exercise",
 		gollm.WithOutput("JSON array of benefits"),
@@ -109,7 +119,7 @@ func TestStructuredOutput(t *testing.T) {
 	)
 
 	response, err := llm.Generate(ctx, prompt)
-	assert.NoError(t, err)
+	require.NoError(t, err)
 	assert.Equal(t, expectedJSON, response)
 
 	// Verify that all expected mock calls were made
@@ -122,7 +132,7 @@ func TestJSONSchemaValidation(t *testing.T) {
 	// Create and configure mock provider
 	mockProvider := new(MockProvider)
 	mockProvider.On("Name").Return("mock")
-	mockProvider.On("SupportsJSONSchema").Return(true)
+	mockProvider.On("SupportsStructuredResponse").Return(true)
 	mockProvider.On("SetOption", mock.Anything, mock.Anything).Return()
 	mockProvider.On("SetDefaultOptions", mock.Anything).Return()
 	mockProvider.On("SetLogger", mock.Anything).Return()
@@ -138,27 +148,26 @@ func TestJSONSchemaValidation(t *testing.T) {
 		gollm.SetModel("mock-model"),
 		gollm.SetAPIKey("test-key"),
 	)
-	assert.NoError(t, err)
+	require.NoError(t, err)
 
-	schema := `{
-        "type": "object",
-        "properties": {
-            "name": {"type": "string"},
-            "age": {"type": "integer", "minimum": 18},
-            "interests": {"type": "array", "items": {"type": "string"}}
-        },
-        "required": ["name", "age", "interests"]
-    }`
-
-	prompt := gollm.NewPrompt("Generate a user profile",
-		gollm.WithOutput(schema),
-		gollm.WithSystemPrompt("You are a JSON-only assistant.", gollm.CacheTypeEphemeral),
+	prompt := gollm.NewPrompt(
+		"Generate a user profile",
+		gollm.WithSystemPrompt(
+			"You are data analyst who specializes in generating user data.",
+			gollm.CacheTypeEphemeral,
+		),
 	)
 
-	response, err := llm.Generate(ctx, prompt, gollm.WithJSONSchemaValidation())
-	assert.NoError(t, err)
+	response, err := llm.Generate(ctx, prompt, llm.WithStructuredResponseSchema(UserProfile{}))
+	require.NoError(t, err)
 	assert.Equal(t, expectedJSON, response)
 
 	// Verify that all expected mock calls were made
 	mockProvider.AssertExpectations(t)
+}
+
+type UserProfile struct {
+	Name      string   `json:"name"`
+	Age       int      `json:"age"`
+	Interests []string `json:"interests"`
 }
