@@ -3,33 +3,24 @@ package gollm
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"sync"
 	"time"
 
-	"github.com/teilomillet/gollm/config"
-	"github.com/teilomillet/gollm/llm"
-	"github.com/teilomillet/gollm/providers"
-	"github.com/teilomillet/gollm/utils"
+	"github.com/weave-labs/gollm/config"
+	"github.com/weave-labs/gollm/internal/logging"
+	"github.com/weave-labs/gollm/llm"
+	"github.com/weave-labs/gollm/providers"
 )
 
 // MOAConfig represents the configuration for the Mixture of Agents (MOA) system.
 // MOA is an ensemble learning approach that combines multiple language models
 // to produce higher quality outputs through iterative refinement and aggregation.
 type MOAConfig struct {
-	// Iterations specifies the number of refinement cycles to perform
-	Iterations int
-
-	// Models defines the set of language models to use in the mixture
-	// Each model is configured using ConfigOption functions
-	Models []ConfigOption
-
-	// MaxParallel limits the number of concurrent model executions
-	// If set to 0, all models run concurrently without limitation
-	MaxParallel int
-
-	// AgentTimeout specifies the maximum duration for each agent's execution
-	// If set to 0, no timeout is applied
+	Models       []ConfigOption
+	Iterations   int
+	MaxParallel  int
 	AgentTimeout time.Duration
 }
 
@@ -45,14 +36,9 @@ type MOALayer struct {
 // through multiple layers and iterations, then aggregates the results using a
 // designated aggregator model.
 type MOA struct {
-	// Config holds the MOA system configuration
-	Config MOAConfig
-
-	// Layers contains the processing layers, each with one or more models
-	Layers []MOALayer
-
-	// Aggregator is the LLM instance used to synthesize outputs from multiple models
 	Aggregator llm.LLM
+	Layers     []MOALayer
+	Config     MOAConfig
 }
 
 // NewMOA creates a new Mixture of Agents instance with the specified configuration
@@ -75,11 +61,11 @@ type MOA struct {
 //	})
 func NewMOA(moaConfig MOAConfig, aggregatorOpts ...ConfigOption) (*MOA, error) {
 	if len(moaConfig.Models) == 0 {
-		return nil, fmt.Errorf("invalid model configuration: at least one model must be specified")
+		return nil, errors.New("invalid model configuration: at least one model must be specified")
 	}
 
 	registry := providers.NewProviderRegistry()
-	logger := utils.NewLogger(utils.LogLevelInfo)
+	logger := logging.NewLogger(logging.LogLevelInfo)
 
 	moa := &MOA{
 		Config: moaConfig,
@@ -131,10 +117,10 @@ func NewMOA(moaConfig MOAConfig, aggregatorOpts ...ConfigOption) (*MOA, error) {
 // 3. This process repeats for the specified number of iterations
 // 4. Results from all iterations are aggregated into the final output
 func (moa *MOA) Generate(ctx context.Context, input string) (string, error) {
-	var layerOutputs []string
+	layerOutputs := make([]string, 0, moa.Config.Iterations)
 
 	// Process through each iteration
-	for i := 0; i < moa.Config.Iterations; i++ {
+	for range moa.Config.Iterations {
 		layerInput := input
 		// Process each layer
 		for _, layer := range moa.Layers {
@@ -169,7 +155,7 @@ func (moa *MOA) Generate(ctx context.Context, input string) (string, error) {
 //   - Combines outputs from all models in the layer
 func (moa *MOA) processLayer(ctx context.Context, layer MOALayer, input string) (string, error) {
 	results := make([]string, len(layer.Models))
-	errors := make([]error, len(layer.Models))
+	errs := make([]error, len(layer.Models))
 
 	// Create a worker pool if parallel processing is enabled
 	var wg sync.WaitGroup
@@ -188,25 +174,26 @@ func (moa *MOA) processLayer(ctx context.Context, layer MOALayer, input string) 
 			}
 
 			// Create a context with timeout if AgentTimeout is set
-			var cancel context.CancelFunc
+			ctxToUse := ctx
 			if moa.Config.AgentTimeout > 0 {
-				ctx, cancel = context.WithTimeout(ctx, moa.Config.AgentTimeout)
+				var cancel context.CancelFunc
+				ctxToUse, cancel = context.WithTimeout(ctx, moa.Config.AgentTimeout)
 				defer cancel()
 			}
 
-			output, err := llmInstance.Generate(ctx, llm.NewPrompt(input))
+			output, err := llmInstance.Generate(ctxToUse, llm.NewPrompt(input))
 			if err != nil {
-				errors[index] = err
+				errs[index] = err
 				return
 			}
-			results[index] = output
+			results[index] = output.AsText()
 		}(i, model)
 	}
 
 	wg.Wait()
 
 	// Check for errors
-	for _, err := range errors {
+	for _, err := range errs {
 		if err != nil {
 			return "", fmt.Errorf("error in layer processing: %w", err)
 		}
@@ -224,9 +211,9 @@ func (moa *MOA) processLayer(ctx context.Context, layer MOALayer, input string) 
 // Returns:
 //   - A combined string containing all model outputs
 func (moa *MOA) combineResults(results []string) string {
-	combined := ""
+	var combined string
 	for _, result := range results {
-		combined += result + "\n---\n"
+		combined += "\n---\n" + result
 	}
 	return combined
 }
@@ -242,8 +229,11 @@ func (moa *MOA) combineResults(results []string) string {
 //   - The final synthesized output
 //   - Any error encountered during aggregation
 func (moa *MOA) aggregate(ctx context.Context, outputs []string) (string, error) {
-	aggregationPrompt := fmt.Sprintf("Synthesise these responses into a single, high-quality response:\n\n%s", moa.combineResults(outputs))
-	response, err := moa.Aggregator.Generate(ctx, llm.NewPrompt(aggregationPrompt))
+	response, err := moa.Aggregator.Generate(ctx, llm.NewPrompt(
+		"Synthesise these responses into a single, high-quality response:\n\n"+moa.combineResults(outputs)))
+	if err != nil {
+		return "", fmt.Errorf("error during aggregation: %w", err)
+	}
 
-	return response, err
+	return response.AsText(), err
 }
