@@ -1,42 +1,44 @@
-// Package presets provides utilities for enhancing Language Learning Model interactions
+// Package presets provide utilities for enhancing Language Learning Model interactions
 // with specific reasoning patterns and comparison capabilities.
 package presets
 
 import (
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
+	"log/slog"
 	"strings"
 
-	"github.com/teilomillet/gollm/config"
-	"github.com/teilomillet/gollm/llm"
-	"github.com/teilomillet/gollm/providers"
-	"github.com/teilomillet/gollm/utils"
+	"github.com/weave-labs/gollm/config"
+	"github.com/weave-labs/gollm/internal/logging"
+	"github.com/weave-labs/gollm/llm"
+	"github.com/weave-labs/gollm/providers"
 )
 
 // ComparisonResult represents the outcome of a model comparison for a specific provider.
 // It is a generic type that can hold any structured response data along with metadata
 // about the generation attempt.
 type ComparisonResult[T any] struct {
-	Provider string // The LLM provider (e.g., "openai", "anthropic")
-	Model    string // The specific model used (e.g., "gpt-4", "claude-2")
-	Response string // Raw response from the model
-	Data     T      // Parsed and validated response data
-	Error    error  // Any error encountered during generation or validation
-	Attempts int    // Number of attempts made to get a valid response
+	Data     T
+	Error    error
+	Response *providers.Response
+	Provider string
+	Model    string
+	Attempts int
 }
 
 // debugLog outputs debug information when debug logging is enabled in the config.
 // It helps track the comparison process and troubleshoot issues.
-func debugLog(config *config.Config, format string, args ...interface{}) {
-	if config.LogLevel == utils.LogLevelDebug {
-		fmt.Printf("[DEBUG] "+format+"\n", args...)
+func debugLog(cfg *config.Config, format string, args ...any) {
+	if cfg.LogLevel == logging.LogLevelDebug {
+		slog.Debug(fmt.Sprintf(format, args...))
 	}
 }
 
 // cleanResponse processes the raw model response to extract valid JSON.
-// It handles common response formats including:
-// - Responses wrapped in markdown code blocks
+// It handles common response formats including
+// - Responses wrapped in Markdown code blocks
 // - Responses with additional text before/after JSON
 // - Responses with multiple JSON objects
 func cleanResponse(response string) string {
@@ -125,98 +127,195 @@ type ValidateFunc[T any] func(T) error
 //	}
 //
 //	// Compare model outputs
-//	results, err := CompareModels[ComplexPerson](
-//	    ctx, prompt, validatePerson, configs...)
 //
-//	// Analyze and display results
-//	analysis := AnalyzeComparisonResults(results)
-//	fmt.Println(analysis)
-func CompareModels[T any](ctx context.Context, prompt string, validateFunc ValidateFunc[T], configs ...*config.Config) ([]ComparisonResult[T], error) {
-	// Validate inputs
+// validateCompareInputs validates the inputs for CompareModels
+func validateCompareInputs[T any](prompt string, validateFunc ValidateFunc[T], configs []*config.Config) error {
 	if prompt == "" {
-		return nil, fmt.Errorf("prompt cannot be empty")
+		return errors.New("prompt cannot be empty")
 	}
 	if validateFunc == nil {
-		return nil, fmt.Errorf("validator function cannot be nil")
+		return errors.New("validator function cannot be nil")
 	}
 	if len(configs) == 0 {
-		return nil, fmt.Errorf("at least one config must be provided")
+		return errors.New("at least one config must be provided")
+	}
+	return nil
+}
+
+// processConfig processes a single configuration and returns whether it succeeded
+func processConfig[T any](
+	ctx context.Context,
+	prompt string,
+	validateFunc ValidateFunc[T],
+	remainingConfig *config.Config,
+	configs []*config.Config,
+	results []ComparisonResult[T],
+	attempt int,
+	logger logging.Logger,
+) (bool, error) {
+	debugLog(
+		remainingConfig,
+		"Attempting generation for %s %s (Attempt %d)",
+		remainingConfig.Provider,
+		remainingConfig.Model,
+		attempt,
+	)
+
+	// Generate response
+	response, err := generateResponse(ctx, prompt, remainingConfig, logger)
+	if err != nil {
+		return handleGenerationError(err, remainingConfig, attempt)
+	}
+
+	// Store result
+	index := findConfigIndex(configs, remainingConfig)
+	updateResult(&results[index], remainingConfig, response, attempt)
+
+	// Process and validate response
+	return processAndValidateResponse[T](&results[index], response, validateFunc, remainingConfig, attempt)
+}
+
+// generateResponse generates a response using the LLM
+func generateResponse(
+	ctx context.Context,
+	prompt string,
+	cfg *config.Config,
+	logger logging.Logger,
+) (*providers.Response, error) {
+	registry := providers.NewProviderRegistry()
+	llmInstance, err := llm.NewLLM(cfg, logger, registry)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create LLM for %s: %w", cfg.Provider, err)
+	}
+	resp, err := llmInstance.Generate(ctx, llm.NewPrompt(prompt))
+	if err != nil {
+		return nil, fmt.Errorf("failed to generate response: %w", err)
+	}
+	return resp, nil
+}
+
+// handleGenerationError handles errors from response generation
+func handleGenerationError(err error, cfg *config.Config, attempt int) (bool, error) {
+	debugLog(cfg, "Error generating response: %v", err)
+
+	// Immediately propagate API errors (like invalid keys)
+	if strings.Contains(err.Error(), "API error") {
+		return false, fmt.Errorf(
+			"API error for %s %s: %w",
+			cfg.Provider,
+			cfg.Model,
+			err,
+		)
+	}
+
+	if attempt == MaxRetryAttempts {
+		return false, fmt.Errorf("failed to generate response after all attempts: %w", err)
+	}
+
+	return false, nil // Will retry
+}
+
+// updateResult updates the comparison result with response data
+func updateResult[T any](
+	result *ComparisonResult[T],
+	cfg *config.Config,
+	response *providers.Response,
+	attempt int,
+) {
+	result.Provider = cfg.Provider
+	result.Model = cfg.Model
+	result.Response = response
+	result.Attempts = attempt
+}
+
+// processAndValidateResponse processes the response and validates it
+func processAndValidateResponse[T any](
+	result *ComparisonResult[T],
+	response *providers.Response,
+	validateFunc ValidateFunc[T],
+	cfg *config.Config,
+	attempt int,
+) (bool, error) {
+	debugLog(cfg, "Raw response received: %v", response)
+
+	cleanedResponse := cleanResponse(response.AsText())
+	debugLog(cfg, "Cleaned response: %s", cleanedResponse)
+
+	result.Response.Content = providers.Text{Value: cleanedResponse}
+
+	// Parse JSON
+	var data T
+	if err := json.Unmarshal([]byte(cleanedResponse), &data); err != nil {
+		debugLog(cfg, "Invalid JSON: %v", err)
+		result.Error = fmt.Errorf("invalid JSON: %w", err)
+		if attempt == MaxRetryAttempts {
+			return false, fmt.Errorf("failed to parse JSON after all attempts: %w", err)
+		}
+		return false, nil // Will retry
+	}
+
+	// Validate data
+	if err := validateFunc(data); err != nil {
+		debugLog(cfg, "Validation failed: %v", err)
+		result.Error = fmt.Errorf("validation failed: %w", err)
+		if attempt == MaxRetryAttempts {
+			return false, fmt.Errorf("validation failed after all attempts: %w", err)
+		}
+		return false, nil // Will retry
+	}
+
+	result.Data = data
+	return true, nil // Success
+}
+
+// results, err := CompareModels[ComplexPerson](
+//
+//	ctx, prompt, validatePerson, configs...)
+//
+// // Analyze and display results
+// analysis := AnalyzeComparisonResults(results)
+// fmt.Println(analysis)
+func CompareModels[T any](
+	ctx context.Context,
+	prompt string,
+	validateFunc ValidateFunc[T],
+	configs ...*config.Config,
+) ([]ComparisonResult[T], error) {
+	// Validate inputs
+	if err := validateCompareInputs(prompt, validateFunc, configs); err != nil {
+		return nil, err
 	}
 
 	results := make([]ComparisonResult[T], len(configs))
 	remainingConfigs := make([]*config.Config, len(configs))
 	copy(remainingConfigs, configs)
 
-	logger := utils.NewLogger(utils.LogLevelDebug)
+	logger := logging.NewLogger(logging.LogLevelDebug)
 
-	for attempt := 1; attempt <= 3; attempt++ {
+	for attempt := 1; attempt <= MaxRetryAttempts; attempt++ {
 		if len(remainingConfigs) == 0 {
 			break
 		}
 
-		newRemainingConfigs := []*config.Config{}
+		var newRemainingConfigs []*config.Config
 
-		for _, config := range remainingConfigs {
-			debugLog(config, "Attempting generation for %s %s (Attempt %d)", config.Provider, config.Model, attempt)
-
-			registry := providers.NewProviderRegistry()
-			llmInstance, err := llm.NewLLM(config, logger, registry)
+		for _, remainingConfig := range remainingConfigs {
+			success, err := processConfig[T](
+				ctx,
+				prompt,
+				validateFunc,
+				remainingConfig,
+				configs,
+				results,
+				attempt,
+				logger,
+			)
 			if err != nil {
-				return nil, fmt.Errorf("failed to create LLM for %s: %w", config.Provider, err)
+				return nil, err
 			}
-
-			response, err := llmInstance.Generate(ctx, llm.NewPrompt(prompt))
-
-			index := findConfigIndex(configs, config)
-			results[index].Provider = config.Provider
-			results[index].Model = config.Model
-			results[index].Response = response
-			results[index].Error = err
-			results[index].Attempts = attempt
-
-			if err != nil {
-				debugLog(config, "Error generating response: %v", err)
-				// Immediately propagate API errors (like invalid keys)
-				if strings.Contains(err.Error(), "API error") {
-					return nil, fmt.Errorf("API error for %s %s: %w", config.Provider, config.Model, err)
-				}
-				if attempt == 3 {
-					return nil, fmt.Errorf("failed to generate response after all attempts: %w", err)
-				}
-				newRemainingConfigs = append(newRemainingConfigs, config)
-				continue
+			if !success {
+				newRemainingConfigs = append(newRemainingConfigs, remainingConfig)
 			}
-
-			debugLog(config, "Raw response received: %s", response)
-
-			cleanedResponse := cleanResponse(response)
-			debugLog(config, "Cleaned response: %s", cleanedResponse)
-
-			results[index].Response = cleanedResponse
-
-			var data T
-			if err := json.Unmarshal([]byte(cleanedResponse), &data); err != nil {
-				debugLog(config, "Invalid JSON: %v", err)
-				results[index].Error = fmt.Errorf("invalid JSON: %w", err)
-				if attempt == 3 {
-					return nil, fmt.Errorf("failed to parse JSON after all attempts: %w", err)
-				}
-				newRemainingConfigs = append(newRemainingConfigs, config)
-				continue
-			}
-
-			if err := validateFunc(data); err != nil {
-				debugLog(config, "Validation failed: %v", err)
-				results[index].Error = fmt.Errorf("validation failed: %w", err)
-				if attempt == 3 {
-					return nil, fmt.Errorf("validation failed after all attempts: %w", err)
-				}
-				newRemainingConfigs = append(newRemainingConfigs, config)
-				continue
-			}
-
-			results[index].Data = data
-			debugLog(config, "Valid response received for %s %s", config.Provider, config.Model)
 		}
 
 		remainingConfigs = newRemainingConfigs
@@ -228,8 +327,8 @@ func CompareModels[T any](ctx context.Context, prompt string, validateFunc Valid
 // findConfigIndex finds the index of a config in the original config slice.
 // This helps maintain result ordering consistent with input configs.
 func findConfigIndex(configs []*config.Config, target *config.Config) int {
-	for i, config := range configs {
-		if config.Provider == target.Provider && config.Model == target.Model {
+	for i, llmConfig := range configs {
+		if llmConfig.Provider == target.Provider && llmConfig.Model == target.Model {
 			return i
 		}
 	}
@@ -295,7 +394,7 @@ func AnalyzeComparisonResults[T any](results []ComparisonResult[T]) string {
 	var analysis strings.Builder
 
 	for _, result := range results {
-		analysis.WriteString(strings.Repeat("-", 40) + "\n")
+		analysis.WriteString(strings.Repeat("-", SeparatorLineLength) + "\n")
 		analysis.WriteString(fmt.Sprintf("Provider: %s, Model: %s\n", result.Provider, result.Model))
 		analysis.WriteString(fmt.Sprintf("Attempts: %d\n", result.Attempts))
 		if result.Error != nil {
@@ -304,12 +403,12 @@ func AnalyzeComparisonResults[T any](results []ComparisonResult[T]) string {
 			prettyJSON, err := json.MarshalIndent(result.Data, "", "  ")
 			if err != nil {
 				analysis.WriteString(fmt.Sprintf("Error prettifying JSON: %v\n", err))
-				analysis.WriteString(fmt.Sprintf("Raw response: %s\n", result.Response))
+				analysis.WriteString(fmt.Sprintf("Raw response: %s\n", result.Response.AsText()))
 			} else {
 				analysis.WriteString(fmt.Sprintf("Response: %s\n", string(prettyJSON)))
 			}
 		}
-		analysis.WriteString(strings.Repeat("-", 40) + "\n")
+		analysis.WriteString(strings.Repeat("-", SeparatorLineLength) + "\n")
 	}
 
 	return analysis.String()
