@@ -6,36 +6,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/teilomillet/gollm"
-)
-
-// QuestionAnswerTemplate defines a structured prompt template for question answering.
-// It guides the LLM to provide clear, concise, and relevant answers while maintaining
-// a consistent output format.
-//
-// The template includes:
-// - Clear question presentation
-// - Directive for concise answers
-// - Structured output format
-//
-// Example generated prompt:
-//
-//	Answer the following question:
-//	
-//	What are the main challenges in quantum computing?
-//	
-//	Directives:
-//	- Provide a clear and concise answer
-//	
-//	Answer:
-var QuestionAnswerTemplate = gollm.NewPromptTemplate(
-	"QuestionAnswer",
-	"Answer the given question",
-	"Answer the following question:\n\n{{.Question}}",
-	gollm.WithPromptOptions(
-		gollm.WithDirectives("Provide a clear and concise answer"),
-		gollm.WithOutput("Answer:"),
-	),
+	"github.com/weave-labs/gollm"
 )
 
 // QuestionAnswer performs question answering with support for context, examples,
@@ -70,9 +41,9 @@ var QuestionAnswerTemplate = gollm.NewPromptTemplate(
 //
 // Example usage with context and examples:
 //
-//	contextInfo := `Quantum computing is an emerging field that uses 
-//	quantum-mechanical phenomena such as superposition and entanglement 
-//	to perform computation. It has the potential to solve certain problems 
+//	contextInfo := `Quantum computing is an emerging field that uses
+//	quantum-mechanical phenomena such as superposition and entanglement
+//	to perform computation. It has the potential to solve certain problems
 //	much faster than classical computers.`
 //
 //	answer, err := QuestionAnswer(ctx, llm,
@@ -89,24 +60,34 @@ var QuestionAnswerTemplate = gollm.NewPromptTemplate(
 // Example response:
 //
 //	The main challenges in quantum computing include:
-//	
+//
 //	1. Decoherence: Quantum states are extremely fragile and can collapse
 //	   due to environmental interactions. This requires sophisticated error
 //	   correction techniques.
-//	
+//
 //	2. Scalability: Building large-scale quantum computers while maintaining
 //	   coherence is technically challenging.
-//	
+//
 //	3. Error Rates: Current quantum gates have relatively high error rates,
 //	   making reliable computations difficult.
-//	
+//
 //	4. Cost and Complexity: Quantum computers require extremely precise
 //	   control systems and specialized operating conditions.
 func QuestionAnswer(ctx context.Context, l gollm.LLM, question string, opts ...gollm.PromptOption) (string, error) {
+	questionAnswerTemplate := gollm.NewPromptTemplate(
+		"QuestionAnswer",
+		"Answer the given question",
+		"Answer the following question:\n\n{{.Question}}",
+		gollm.WithPromptOptions(
+			gollm.WithDirectives("Provide a clear and concise answer"),
+			gollm.WithOutput("Answer:"),
+		),
+	)
+
 	if ctx == nil {
 		ctx = context.Background()
 	}
-	prompt, err := QuestionAnswerTemplate.Execute(map[string]interface{}{
+	prompt, err := questionAnswerTemplate.Execute(map[string]any{
 		"Question": question,
 	})
 	if err != nil {
@@ -117,5 +98,5 @@ func QuestionAnswer(ctx context.Context, l gollm.LLM, question string, opts ...g
 	if err != nil {
 		return "", fmt.Errorf("failed to generate response: %w", err)
 	}
-	return response, nil
+	return response.AsText(), nil
 }
