@@ -3,50 +3,61 @@ package providers
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
+	"slices"
 
-	"github.com/teilomillet/gollm/config"
-	"github.com/teilomillet/gollm/types"
-	"github.com/teilomillet/gollm/utils"
+	"github.com/weave-labs/gollm/config"
+	"github.com/weave-labs/gollm/internal/logging"
+)
+
+// Groq-specific parameter keys
+const (
+	groqKeyMessages                 = "messages"
+	groqKeySystemPrompt             = "system_prompt"
+	groqKeyTools                    = "tools"
+	groqKeyToolChoice               = "tool_choice"
+	groqKeyStructuredMessages       = "structured_messages"
+	groqKeyStructuredResponseSchema = "structured_response_schema"
+	groqKeyMaxTokens                = "max_tokens"
+	groqKeyStream                   = "stream"
+	groqKeyModel                    = "model"
 )
 
 // GroqProvider implements the Provider interface for Groq's API.
 // It supports Groq's optimized language models and provides access to their
 // high-performance inference capabilities.
 type GroqProvider struct {
-	apiKey       string                 // API key for authentication
-	model        string                 // Model identifier (e.g., "llama2-70b", "mixtral-8x7b")
-	extraHeaders map[string]string      // Additional HTTP headers
-	options      map[string]interface{} // Model-specific options
-	logger       utils.Logger           // Logger instance
+	logger       logging.Logger
+	extraHeaders map[string]string
+	options      map[string]any
+	apiKey       string
+	model        string
 }
 
 // NewGroqProvider creates a new Groq provider instance.
 // It initializes the provider with the given API key, model, and optional headers.
-//
-// Parameters:
-//   - apiKey: Groq API key for authentication
-//   - model: The model to use (e.g., "llama2-70b", "mixtral-8x7b")
-//   - extraHeaders: Additional HTTP headers for requests
-//
-// Returns:
-//   - A configured Groq Provider instance
-func NewGroqProvider(apiKey, model string, extraHeaders map[string]string) Provider {
+func NewGroqProvider(apiKey, model string, extraHeaders map[string]string) *GroqProvider {
 	if extraHeaders == nil {
 		extraHeaders = make(map[string]string)
 	}
-	return &GroqProvider{
+
+	p := &GroqProvider{
 		apiKey:       apiKey,
 		model:        model,
 		extraHeaders: extraHeaders,
-		options:      make(map[string]interface{}),
-		logger:       utils.NewLogger(utils.LogLevelInfo),
+		options:      make(map[string]any),
+		logger:       logging.NewLogger(logging.LogLevelInfo),
 	}
+
+	// Register capabilities based on model
+	p.registerCapabilities()
+	return p
 }
 
 // SetLogger configures the logger for the Groq provider.
 // This is used for debugging and monitoring API interactions.
-func (p *GroqProvider) SetLogger(logger utils.Logger) {
+func (p *GroqProvider) SetLogger(logger logging.Logger) {
 	p.logger = logger
 }
 
@@ -55,36 +66,92 @@ func (p *GroqProvider) Name() string {
 	return "groq"
 }
 
+// registerCapabilities registers capabilities for all known Groq models
+func (p *GroqProvider) registerCapabilities() {
+	registry := GetRegistry()
+
+	// Define all known Groq models
+	allModels := []string{
+		// Llama models
+		"llama-3.1-405b-reasoning",
+		"llama-3.1-70b-versatile",
+		"llama-3.1-8b-instant",
+		"llama3-groq-70b-8192-tool-use-preview",
+		"llama3-groq-8b-8192-tool-use-preview",
+		"llama-3.2-1b-preview",
+		"llama-3.2-3b-preview",
+		"llama-3.2-11b-text-preview",
+		"llama-3.2-90b-text-preview",
+		"llama-guard-3-8b",
+		"llama3-70b-8192",
+		"llama3-8b-8192",
+
+		// Mixtral models
+		"mixtral-8x7b-32768",
+
+		// Gemma models
+		"gemma-7b-it",
+		"gemma2-9b-it",
+
+		// DeepSeek models
+		"deepseek-r1-distill-llama-70b",
+
+		// OpenAI models on Groq
+		"openai/gpt-oss-20b",
+		"openai/gpt-oss-120b",
+
+		// Moonshot models
+		"moonshotai/kimi-k2-instruct",
+
+		// Meta Llama 4 models
+		"meta-llama/llama-4-maverick-17b-128e-instruct",
+		"meta-llama/llama-4-scout-17b-16e-instruct",
+	}
+
+	for _, model := range allModels {
+		// Structured response support - all Groq models
+		registry.Register(ProviderGroq, model, CapStructuredResponse, StructuredResponseConfig{
+			RequiresToolUse:  false,
+			MaxSchemaDepth:   10,
+			SupportedFormats: []string{"json"},
+			RequiresJSONMode: true,
+		})
+
+		// Function calling - most Groq models support it (exclude guard models)
+		if !slices.Contains([]string{"llama-guard-3-8b"}, model) {
+			registry.Register(ProviderGroq, model, CapFunctionCalling, FunctionCallingConfig{
+				MaxFunctions:      100,
+				SupportsParallel:  true,
+				MaxParallelCalls:  10,
+				SupportsStreaming: true,
+			})
+		}
+
+		// Streaming - most models support it
+		if !slices.Contains([]string{"llama-guard-3-8b"}, model) {
+			registry.Register(ProviderGroq, model, CapStreaming, StreamingConfig{
+				SupportsSSE:    true,
+				BufferSize:     4096,
+				ChunkDelimiter: "data: ",
+				SupportsUsage:  false,
+			})
+		}
+	}
+}
+
+// HasCapability checks if a capability is supported
+func (p *GroqProvider) HasCapability(capability Capability, model string) bool {
+	targetModel := p.model
+	if model != "" {
+		targetModel = model
+	}
+	return GetRegistry().HasCapability(ProviderGroq, targetModel, capability)
+}
+
 // Endpoint returns the Groq API endpoint URL.
 // This is "https://api.groq.com/openai/v1/chat/completions".
 func (p *GroqProvider) Endpoint() string {
 	return "https://api.groq.com/openai/v1/chat/completions"
-}
-
-// SetOption sets a model-specific option for the Groq provider.
-// Supported options include:
-//   - temperature: Controls randomness (0.0 to 1.0)
-//   - max_tokens: Maximum tokens in the response
-//   - top_p: Nucleus sampling parameter
-//   - top_k: Top-k sampling parameter
-func (p *GroqProvider) SetOption(key string, value interface{}) {
-	p.options[key] = value
-}
-
-// SetDefaultOptions configures standard options from the global configuration.
-// This includes temperature, max tokens, and sampling parameters.
-func (p *GroqProvider) SetDefaultOptions(config *config.Config) {
-	p.SetOption("temperature", config.Temperature)
-	p.SetOption("max_tokens", config.MaxTokens)
-	if config.Seed != nil {
-		p.SetOption("seed", *config.Seed)
-	}
-}
-
-// SupportsJSONSchema indicates whether this provider supports JSON schema validation.
-// Currently, Groq does not natively support JSON schema validation.
-func (p *GroqProvider) SupportsJSONSchema() bool {
-	return false
 }
 
 // Headers returns the HTTP headers required for Groq API requests.
@@ -102,76 +169,96 @@ func (p *GroqProvider) Headers() map[string]string {
 	return headers
 }
 
-// PrepareRequest creates the request body for a Groq API call.
-// It formats the prompt and options according to Groq's API requirements.
-//
-// Parameters:
-//   - prompt: The input text or conversation
-//   - options: Additional parameters for the request
-//
-// Returns:
-//   - Serialized JSON request body
-//   - Any error encountered during preparation
-func (p *GroqProvider) PrepareRequest(prompt string, options map[string]interface{}) ([]byte, error) {
-	requestBody := map[string]interface{}{
-		"model": p.model,
-		"messages": []map[string]string{
-			{"role": "user", "content": prompt},
-		},
-	}
-
-	// First, add the default options
-	for k, v := range p.options {
-		requestBody[k] = v
-	}
-
-	// Then, add any additional options (which may override defaults)
-	for k, v := range options {
-		requestBody[k] = v
-	}
-
-	return json.Marshal(requestBody)
+// SetExtraHeaders configures additional HTTP headers for API requests.
+// This allows for custom headers needed for specific features or requirements.
+func (p *GroqProvider) SetExtraHeaders(extraHeaders map[string]string) {
+	p.extraHeaders = extraHeaders
 }
 
-// PrepareRequestWithSchema creates a request with JSON schema validation.
-// Since Groq doesn't support schema validation natively, this falls back to
-// standard request preparation.
-func (p *GroqProvider) PrepareRequestWithSchema(prompt string, options map[string]interface{}, schema interface{}) ([]byte, error) {
-	requestBody := map[string]interface{}{
-		"model": p.model,
-		"messages": []map[string]string{
-			{"role": "user", "content": prompt},
-		},
-		"response_format": map[string]interface{}{
-			"type":   "json_schema",
-			"schema": schema,
-		},
+// SetDefaultOptions configures standard options from the global configuration.
+// This includes temperature, max tokens, and sampling parameters.
+func (p *GroqProvider) SetDefaultOptions(cfg *config.Config) {
+	p.SetOption("temperature", cfg.Temperature)
+	p.SetOption(groqKeyMaxTokens, cfg.MaxTokens)
+	if cfg.Seed != nil {
+		p.SetOption("seed", *cfg.Seed)
+	}
+}
+
+// SetOption sets a model-specific option for the Groq provider.
+// Supported options include:
+//   - temperature: Controls randomness (0.0 to 1.0)
+//   - max_tokens: Maximum tokens in the response
+//   - top_p: Nucleus sampling parameter
+//   - top_k: Top-k sampling parameter
+func (p *GroqProvider) SetOption(key string, value any) {
+	p.options[key] = value
+}
+
+// PrepareRequest creates the request body for a Groq API call using the new Request structure.
+// It formats the messages and options according to Groq's API requirements.
+func (p *GroqProvider) PrepareRequest(req *Request, options map[string]any) ([]byte, error) {
+	// Determine which model to use
+	model := p.model
+	if req.Model != "" {
+		model = req.Model
+	} else if m, ok := options["model"].(string); ok && m != "" {
+		model = m
 	}
 
-	// Add any additional options
-	for k, v := range options {
-		requestBody[k] = v
+	requestBody := p.initializeRequestBodyWithModel(model)
+
+	p.addMessagesToRequestBody(requestBody, req.Messages, options)
+
+	if req.SystemPrompt != "" {
+		p.addSystemPromptToRequestBody(requestBody, req.SystemPrompt)
 	}
 
-	// Add strict option if provided
-	if strict, ok := options["strict"].(bool); ok && strict {
-		requestBody["response_format"].(map[string]interface{})["strict"] = true
+	if req.ResponseSchema != nil && options[groqKeyStream] != true {
+		p.addStructuredResponseToRequest(requestBody, req.ResponseSchema)
 	}
 
-	return json.Marshal(requestBody)
+	p.addRemainingOptions(requestBody, options)
+
+	data, err := json.Marshal(requestBody)
+	if err != nil {
+		return nil, fmt.Errorf("failed to marshal request body: %w", err)
+	}
+	return data, nil
+}
+
+// PrepareStreamRequest creates the request body for a streaming Groq API call.
+// It uses the same structure as PrepareRequest but adds the stream parameter.
+func (p *GroqProvider) PrepareStreamRequest(req *Request, options map[string]any) ([]byte, error) {
+	// Determine which model to use
+	model := p.model
+	if req.Model != "" {
+		model = req.Model
+	} else if m, ok := options["model"].(string); ok && m != "" {
+		model = m
+	}
+
+	if !p.HasCapability(CapStreaming, model) {
+		return nil, errors.New("streaming is not supported by this provider")
+	}
+
+	if options == nil {
+		options = make(map[string]any)
+	}
+	options[groqKeyStream] = true
+
+	return p.PrepareRequest(req, options)
 }
 
 // ParseResponse extracts the generated text from the Groq API response.
 // It handles Groq's response format and extracts the content.
-//
-// Parameters:
-//   - body: Raw API response body
-//
-// Returns:
-//   - Generated text content
-//   - Any error encountered during parsing
-func (p *GroqProvider) ParseResponse(body []byte) (string, error) {
+func (p *GroqProvider) ParseResponse(body []byte) (*Response, error) {
 	var response struct {
+		Usage *struct {
+			PromptTokens     int64 `json:"prompt_tokens"`
+			CompletionTokens int64 `json:"completion_tokens"`
+			TotalTokens      int64 `json:"total_tokens"`
+		} `json:"usage"`
 		Choices []struct {
 			Message struct {
 				Content string `json:"content"`
@@ -181,51 +268,22 @@ func (p *GroqProvider) ParseResponse(body []byte) (string, error) {
 
 	err := json.Unmarshal(body, &response)
 	if err != nil {
-		return "", fmt.Errorf("error parsing response: %w", err)
+		return nil, fmt.Errorf("error parsing response: %w", err)
 	}
 
 	if len(response.Choices) == 0 || response.Choices[0].Message.Content == "" {
-		return "", fmt.Errorf("empty response from API")
+		return nil, errors.New("empty response from API")
 	}
 
-	return response.Choices[0].Message.Content, nil
-}
-
-// HandleFunctionCalls processes function calling capabilities.
-// Since Groq doesn't support function calling natively, this returns nil.
-func (p *GroqProvider) HandleFunctionCalls(body []byte) ([]byte, error) {
-	response := string(body)
-	functionCalls, err := utils.ExtractFunctionCalls(response)
-	if err != nil {
-		return nil, fmt.Errorf("error extracting function calls: %w", err)
+	resp := &Response{Content: Text{Value: response.Choices[0].Message.Content}}
+	if response.Usage != nil {
+		resp.Usage = NewUsage(response.Usage.PromptTokens, 0, response.Usage.CompletionTokens, 0, 0)
 	}
-
-	if len(functionCalls) == 0 {
-		return nil, nil // No function calls found
-	}
-
-	return json.Marshal(functionCalls)
-}
-
-// SetExtraHeaders configures additional HTTP headers for API requests.
-// This allows for custom headers needed for specific features or requirements.
-func (p *GroqProvider) SetExtraHeaders(extraHeaders map[string]string) {
-	p.extraHeaders = extraHeaders
-}
-
-// SupportsStreaming returns whether the provider supports streaming responses
-func (p *GroqProvider) SupportsStreaming() bool {
-	return true
-}
-
-// PrepareStreamRequest prepares a request body for streaming
-func (p *GroqProvider) PrepareStreamRequest(prompt string, options map[string]interface{}) ([]byte, error) {
-	options["stream"] = true
-	return p.PrepareRequest(prompt, options)
+	return resp, nil
 }
 
 // ParseStreamResponse parses a single chunk from a streaming response
-func (p *GroqProvider) ParseStreamResponse(chunk []byte) (string, error) {
+func (p *GroqProvider) ParseStreamResponse(chunk []byte) (*Response, error) {
 	var response struct {
 		Choices []struct {
 			Delta struct {
@@ -234,49 +292,99 @@ func (p *GroqProvider) ParseStreamResponse(chunk []byte) (string, error) {
 		} `json:"choices"`
 	}
 	if err := json.Unmarshal(chunk, &response); err != nil {
-		return "", err
+		return nil, fmt.Errorf("malformed response: %w", err)
 	}
-	if len(response.Choices) == 0 {
-		return "", nil
+	if len(response.Choices) == 0 || response.Choices[0].Delta.Content == "" {
+		return nil, errors.New("skip resp")
 	}
-	return response.Choices[0].Delta.Content, nil
+	return &Response{Content: Text{Value: response.Choices[0].Delta.Content}}, nil
 }
 
-// PrepareRequestWithMessages creates a request body using structured message objects
-// rather than a flattened prompt string.
-func (p *GroqProvider) PrepareRequestWithMessages(messages []types.MemoryMessage, options map[string]interface{}) ([]byte, error) {
-	request := map[string]interface{}{
-		"model":    p.model,
-		"messages": []map[string]interface{}{},
-	}
+// Private helper methods
 
-	// Add system prompt if present
-	if systemPrompt, ok := options["system_prompt"].(string); ok && systemPrompt != "" {
-		request["messages"] = append(request["messages"].([]map[string]interface{}), map[string]interface{}{
-			"role":    "system",
-			"content": systemPrompt,
-		})
-	}
+// initializeRequestBody creates the base request body with model information
 
-	// Convert structured messages to Groq format (OpenAI compatible)
+// initializeRequestBodyWithModel creates the base request body with specified model
+func (p *GroqProvider) initializeRequestBodyWithModel(model string) map[string]any {
+	return map[string]any{
+		groqKeyModel: model,
+	}
+}
+
+// addMessagesToRequestBody adds messages to the request body in Groq format
+func (p *GroqProvider) addMessagesToRequestBody(
+	requestBody map[string]any,
+	messages []Message,
+	_ map[string]any,
+) {
+	groqMessages := make([]map[string]any, 0, len(messages))
+
 	for _, msg := range messages {
-		request["messages"] = append(request["messages"].([]map[string]interface{}), map[string]interface{}{
+		groqMessage := map[string]any{
 			"role":    msg.Role,
 			"content": msg.Content,
-		})
+		}
+
+		// Add optional fields if present
+		if msg.Name != "" {
+			groqMessage["name"] = msg.Name
+		}
+		if msg.ToolCallID != "" {
+			groqMessage["tool_call_id"] = msg.ToolCallID
+		}
+		if len(msg.ToolCalls) > 0 {
+			groqMessage["tool_calls"] = msg.ToolCalls
+		}
+
+		groqMessages = append(groqMessages, groqMessage)
 	}
 
-	// Add other options from provider and request
+	requestBody[groqKeyMessages] = groqMessages
+}
+
+// addSystemPromptToRequestBody adds system prompt as a system message
+func (p *GroqProvider) addSystemPromptToRequestBody(requestBody map[string]any, systemPrompt string) {
+	if messages, ok := requestBody[groqKeyMessages].([]map[string]any); ok {
+		systemMessage := map[string]any{
+			"role":    "system",
+			"content": systemPrompt,
+		}
+		// Prepend system message
+		requestBody[groqKeyMessages] = append([]map[string]any{systemMessage}, messages...)
+	}
+}
+
+// addStructuredResponseToRequest adds structured response schema to the request
+func (p *GroqProvider) addStructuredResponseToRequest(requestBody map[string]any, schema any) {
+	requestBody["response_format"] = map[string]any{
+		"type":   "json_schema",
+		"schema": schema,
+	}
+}
+
+// addRemainingOptions adds provider options and request options to the request body
+func (p *GroqProvider) addRemainingOptions(requestBody map[string]any, options map[string]any) {
+	// Add provider options first
 	for k, v := range p.options {
-		if k != "messages" {
-			request[k] = v
-		}
-	}
-	for k, v := range options {
-		if k != "messages" && k != "system_prompt" && k != "structured_messages" {
-			request[k] = v
+		if !p.isGlobalOption(k) {
+			requestBody[k] = v
 		}
 	}
 
-	return json.Marshal(request)
+	// Add request options (may override provider options)
+	for k, v := range options {
+		if !p.isGlobalOption(k) {
+			requestBody[k] = v
+		}
+	}
+}
+
+// isGlobalOption checks if a key is a global option that should not be added to request body
+func (p *GroqProvider) isGlobalOption(key string) bool {
+	switch key {
+	case groqKeyMessages, groqKeySystemPrompt, groqKeyStructuredMessages, groqKeyStructuredResponseSchema:
+		return true
+	default:
+		return false
+	}
 }
