@@ -1,71 +1,64 @@
-// Package providers implements LLM provider interfaces and implementations.
 package providers
 
 import (
 	"bytes"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io"
 	"strings"
 
-	"github.com/teilomillet/gollm/config"
-	"github.com/teilomillet/gollm/types"
-	"github.com/teilomillet/gollm/utils"
+	"github.com/weave-labs/gollm/internal/models"
+
+	"github.com/weave-labs/gollm/config"
+	"github.com/weave-labs/gollm/internal/logging"
+)
+
+const (
+	openAIKeyMaxTokens           = "max_tokens"
+	openAIKeyToolChoice          = "tool_choice"
+	openAIKeySystemPrompt        = "system_prompt"
+	openAIKeyTools               = "tools"
+	openAIKeyStructuredMessages  = "structured_messages"
+	openAIKeyMaxCompletionTokens = "max_completion_tokens"
+	openAIKeyStream              = "stream"
 )
 
 // OpenAIProvider implements the Provider interface for OpenAI's API.
 // It supports GPT models and provides access to OpenAI's language model capabilities,
 // including function calling, JSON mode, and structured output validation.
 type OpenAIProvider struct {
-	apiKey       string                 // API key for authentication
-	model        string                 // Model identifier (e.g., "gpt-4", "gpt-4o-mini")
-	extraHeaders map[string]string      // Additional HTTP headers
-	options      map[string]interface{} // Model-specific options
-	logger       utils.Logger           // Logger instance
+	logger       logging.Logger
+	extraHeaders map[string]string
+	options      map[string]any
+	apiKey       string
+	model        string
 }
 
 // NewOpenAIProvider creates a new OpenAI provider instance.
 // It initializes the provider with the given API key, model, and optional headers.
-//
-// Parameters:
-//   - apiKey: OpenAI API key for authentication
-//   - model: The model to use (e.g., "gpt-4", "gpt-3.5-turbo")
-//   - extraHeaders: Additional HTTP headers for requests
-//
-// Returns:
-//   - A configured OpenAI Provider instance
-func NewOpenAIProvider(apiKey, model string, extraHeaders map[string]string) Provider {
+func NewOpenAIProvider(apiKey, model string, extraHeaders map[string]string) *OpenAIProvider {
 	if extraHeaders == nil {
 		extraHeaders = make(map[string]string)
 	}
-	return &OpenAIProvider{
+
+	p := &OpenAIProvider{
 		apiKey:       apiKey,
 		model:        model,
 		extraHeaders: extraHeaders,
-		options:      make(map[string]interface{}),
-		logger:       utils.NewLogger(utils.LogLevelInfo),
+		options:      make(map[string]any),
+		logger:       logging.NewLogger(logging.LogLevelInfo),
 	}
+
+	// Register capabilities with the global registry
+	p.registerCapabilities()
+	return p
 }
 
 // SetLogger configures the logger for the OpenAI provider.
 // This is used for debugging and monitoring API interactions.
-func (p *OpenAIProvider) SetLogger(logger utils.Logger) {
+func (p *OpenAIProvider) SetLogger(logger logging.Logger) {
 	p.logger = logger
-}
-
-// needsMaxCompletionTokens checks if the model requires max_completion_tokens instead of max_tokens
-func (p *OpenAIProvider) needsMaxCompletionTokens() bool {
-	// Check for models that start with "o"
-	if strings.HasPrefix(p.model, "o") {
-		return true
-	}
-
-	// Check for gpt-4o and similar models
-	if strings.Contains(p.model, "4o") || strings.Contains(p.model, "-o") {
-		return true
-	}
-
-	return false
 }
 
 // SetOption sets a specific option for the OpenAI provider.
@@ -76,21 +69,22 @@ func (p *OpenAIProvider) needsMaxCompletionTokens() bool {
 //   - frequency_penalty: Repetition reduction
 //   - presence_penalty: Topic steering
 //   - seed: Deterministic sampling seed
-func (p *OpenAIProvider) SetOption(key string, value interface{}) {
+func (p *OpenAIProvider) SetOption(key string, value any) {
 	// Handle max_tokens conversion for "o" models
-	if key == "max_tokens" {
+	switch key {
+	case openAIKeyMaxTokens:
 		if p.needsMaxCompletionTokens() {
 			// For models requiring max_completion_tokens, use that instead
 			key = "max_completion_tokens"
 			// Delete max_tokens if it was previously set
-			delete(p.options, "max_tokens")
+			delete(p.options, openAIKeyMaxTokens)
 		} else {
 			// For models using max_tokens, make sure max_completion_tokens is not set
 			delete(p.options, "max_completion_tokens")
 		}
-	} else if key == "max_completion_tokens" {
+	case "max_completion_tokens":
 		// If explicitly setting max_completion_tokens, remove max_tokens to avoid conflicts
-		delete(p.options, "max_tokens")
+		delete(p.options, openAIKeyMaxTokens)
 	}
 
 	p.options[key] = value
@@ -99,13 +93,21 @@ func (p *OpenAIProvider) SetOption(key string, value interface{}) {
 
 // SetDefaultOptions configures standard options from the global configuration.
 // This includes temperature, max tokens, and sampling parameters.
-func (p *OpenAIProvider) SetDefaultOptions(config *config.Config) {
-	p.SetOption("temperature", config.Temperature)
-	p.SetOption("max_tokens", config.MaxTokens)
-	if config.Seed != nil {
-		p.SetOption("seed", *config.Seed)
+func (p *OpenAIProvider) SetDefaultOptions(cfg *config.Config) {
+	p.SetOption("temperature", cfg.Temperature)
+	p.SetOption(openAIKeyMaxTokens, cfg.MaxTokens)
+	if cfg.Seed != nil {
+		p.SetOption("seed", *cfg.Seed)
 	}
-	p.logger.Debug("Default options set", "temperature", config.Temperature, "max_tokens", config.MaxTokens, "seed", config.Seed)
+	p.logger.Debug(
+		"Default options set",
+		"temperature",
+		cfg.Temperature,
+		openAIKeyMaxTokens,
+		cfg.MaxTokens,
+		"seed",
+		cfg.Seed,
+	)
 }
 
 // Name returns "openai" as the provider identifier.
@@ -113,16 +115,128 @@ func (p *OpenAIProvider) Name() string {
 	return "openai"
 }
 
+// registerCapabilities registers capabilities for all known OpenAI models
+func (p *OpenAIProvider) registerCapabilities() {
+	registry := GetRegistry()
+
+	// Define all known OpenAI models
+	allModels := []string{
+		// GPT-4o models
+		"gpt-4o", "gpt-4o-mini", "gpt-4o-2024-11-20", "gpt-4o-2024-08-06", "gpt-4o-2024-05-13",
+		"gpt-4o-mini-2024-07-18",
+
+		// GPT-4 Turbo models
+		"gpt-4-turbo", "gpt-4-turbo-2024-04-09", "gpt-4-turbo-preview", "gpt-4-0125-preview",
+		"gpt-4-1106-preview", "gpt-4-turbo-vision-preview",
+
+		// GPT-4 models
+		"gpt-4", "gpt-4-0613", "gpt-4-0314", "gpt-4-vision-preview",
+
+		// GPT-3.5 Turbo models
+		"gpt-3.5-turbo", "gpt-3.5-turbo-0125", "gpt-3.5-turbo-1106", "gpt-3.5-turbo-0613",
+		"gpt-3.5-turbo-16k", "gpt-3.5-turbo-16k-0613",
+
+		// O1 models
+		"o1-preview", "o1-mini", "o1-preview-2024-09-12", "o1-mini-2024-09-12",
+	}
+
+	for _, model := range allModels {
+		// O1 models have limited capabilities
+		if strings.HasPrefix(model, "o1") {
+			// Only register streaming for O1 models
+			registry.Register(ProviderOpenAI, model, CapStreaming, StreamingConfig{
+				SupportsSSE:    true,
+				BufferSize:     4096,
+				ChunkDelimiter: "data: ",
+				SupportsUsage:  false,
+			})
+			continue
+		}
+
+		// GPT-4o models - advanced structured response
+		if strings.HasPrefix(model, "gpt-4o") || strings.HasPrefix(model, "gpt-4-turbo") {
+			registry.Register(ProviderOpenAI, model, CapStructuredResponse, StructuredResponseConfig{
+				RequiresToolUse:  false,
+				MaxSchemaDepth:   15,
+				SupportedFormats: []string{"json", "json_schema"},
+				RequiresJSONMode: true,
+			})
+		} else if strings.HasPrefix(model, "gpt-4") {
+			// Regular GPT-4 models
+			registry.Register(ProviderOpenAI, model, CapStructuredResponse, StructuredResponseConfig{
+				RequiresToolUse:  false,
+				MaxSchemaDepth:   15,
+				SupportedFormats: []string{"json_schema", "json_object"},
+				RequiresJSONMode: true,
+			})
+		} else if strings.HasPrefix(model, "gpt-3.5-turbo") {
+			// GPT-3.5 models - limited structured response support
+			if model == "gpt-3.5-turbo-0125" || model == "gpt-3.5-turbo-1106" {
+				registry.Register(ProviderOpenAI, model, CapStructuredResponse, StructuredResponseConfig{
+					RequiresToolUse:  false,
+					MaxSchemaDepth:   10,
+					SupportedFormats: []string{"json"},
+					RequiresJSONMode: true,
+				})
+			}
+		}
+
+		// Function calling
+		if strings.HasPrefix(model, "gpt-4") {
+			registry.Register(ProviderOpenAI, model, CapFunctionCalling, FunctionCallingConfig{
+				MaxFunctions:      128,
+				SupportsParallel:  true,
+				MaxParallelCalls:  10,
+				SupportsStreaming: true,
+			})
+		} else if strings.HasPrefix(model, "gpt-3.5-turbo") {
+			registry.Register(ProviderOpenAI, model, CapFunctionCalling, FunctionCallingConfig{
+				MaxFunctions:      64,
+				SupportsParallel:  true,
+				MaxParallelCalls:  5,
+				SupportsStreaming: false,
+			})
+		}
+
+		// All OpenAI models support streaming (including O1 which was handled above)
+		if !strings.HasPrefix(model, "o1") {
+			registry.Register(ProviderOpenAI, model, CapStreaming, StreamingConfig{
+				SupportsSSE:    true,
+				BufferSize:     4096,
+				ChunkDelimiter: "data: ",
+				SupportsUsage:  strings.HasPrefix(model, "gpt-4"),
+			})
+		}
+
+		// Vision for specific models
+		visionModels := []string{"gpt-4o", "gpt-4-turbo", "gpt-4-vision"}
+		for _, vm := range visionModels {
+			if strings.HasPrefix(model, vm) {
+				registry.Register(ProviderOpenAI, model, CapVision, VisionConfig{
+					MaxImageSize:        20 * 1024 * 1024,
+					SupportedFormats:    []string{"jpeg", "png", "gif", "webp"},
+					MaxImagesPerRequest: 10,
+					SupportsVideoFrames: strings.Contains(model, "4o"),
+				})
+				break
+			}
+		}
+	}
+}
+
+// HasCapability checks if a capability is supported
+func (p *OpenAIProvider) HasCapability(capability Capability, model string) bool {
+	targetModel := p.model
+	if model != "" {
+		targetModel = model
+	}
+	return GetRegistry().HasCapability(ProviderOpenAI, targetModel, capability)
+}
+
 // Endpoint returns the OpenAI API endpoint URL.
 // For API version 1, this is "https://api.openai.com/v1/chat/completions".
 func (p *OpenAIProvider) Endpoint() string {
 	return "https://api.openai.com/v1/chat/completions"
-}
-
-// SupportsJSONSchema indicates that OpenAI supports native JSON schema validation
-// through its function calling and JSON mode capabilities.
-func (p *OpenAIProvider) SupportsJSONSchema() bool {
-	return true
 }
 
 // Headers returns the required HTTP headers for OpenAI API requests.
@@ -144,326 +258,101 @@ func (p *OpenAIProvider) Headers() map[string]string {
 	return headers
 }
 
-// PrepareRequest creates the request body for an OpenAI API call.
-// It handles:
-//   - Message formatting
-//   - System messages
-//   - Function/tool definitions
-//   - Model-specific options
-//
-// Parameters:
-//   - prompt: The input text or conversation
-//   - options: Additional parameters for the request
-//
-// Returns:
-//   - Serialized JSON request body
-//   - Any error encountered during preparation
-func (p *OpenAIProvider) PrepareRequest(prompt string, options map[string]interface{}) ([]byte, error) {
-	request := map[string]interface{}{
-		"model":    p.model,
-		"messages": []map[string]interface{}{},
+// PrepareRequest creates the request body for an OpenAI API call
+func (p *OpenAIProvider) PrepareRequest(req *Request, options map[string]any) ([]byte, error) {
+	// Determine which model to use
+	model := p.model
+	if req.Model != "" {
+		model = req.Model
+	} else if m, ok := options["model"].(string); ok && m != "" {
+		model = m
 	}
 
-	// Handle system prompt as developer message
-	if systemPrompt, ok := options["system_prompt"].(string); ok && systemPrompt != "" {
-		request["messages"] = append(request["messages"].([]map[string]interface{}), map[string]interface{}{
-			"role":    "developer",
-			"content": systemPrompt,
-		})
+	requestBody := p.initializeOpenAIRequestWithModel(model)
+
+	// Handle system prompt from Request or options
+	systemPrompt := p.extractSystemPromptFromRequest(req, options)
+	if systemPrompt != "" {
+		p.addSystemPromptToRequestBody(requestBody, systemPrompt)
 	}
 
-	// Add user message
-	request["messages"] = append(request["messages"].([]map[string]interface{}), map[string]interface{}{
-		"role":    "user",
-		"content": prompt,
-	})
+	// Add messages from the Request
+	p.addMessagesToRequestBody(requestBody, req.Messages, options)
 
-	// Handle tool_choice
-	if toolChoice, ok := options["tool_choice"].(string); ok {
-		request["tool_choice"] = toolChoice
+	// Handle tools if present in options
+	p.handleToolsForRequest(requestBody, options)
+
+	// Handle structured response schema
+	if req.ResponseSchema != nil && p.HasCapability(CapStructuredResponse, model) {
+		p.addStructuredResponseToRequest(requestBody, req.ResponseSchema)
 	}
 
-	// Handle tools
-	if tools, ok := options["tools"].([]utils.Tool); ok && len(tools) > 0 {
-		openAITools := make([]map[string]interface{}, len(tools))
-		for i, tool := range tools {
-			openAITools[i] = map[string]interface{}{
-				"type": "function",
-				"function": map[string]interface{}{
-					"name":        tool.Function.Name,
-					"description": tool.Function.Description,
-					"parameters":  tool.Function.Parameters,
-				},
-				"strict": true, // Add this if you want strict mode
-			}
-		}
-		request["tools"] = openAITools
-	}
+	// Add remaining options
+	p.addRemainingOptions(requestBody, options)
 
-	// Create a merged copy of options to handle token parameters properly
-	mergedOptions := make(map[string]interface{})
-
-	// First add options from provider (p.options)
-	for k, v := range p.options {
-		if k != "tools" && k != "tool_choice" && k != "system_prompt" {
-			mergedOptions[k] = v
-		}
-	}
-
-	// Then add options from the function parameters (may override provider options)
-	for k, v := range options {
-		if k != "tools" && k != "tool_choice" && k != "system_prompt" {
-			mergedOptions[k] = v
-		}
-	}
-
-	// Handle max_tokens/max_completion_tokens conflict
-	// For models that need max_completion_tokens, ensure we use that and not max_tokens
-	if p.needsMaxCompletionTokens() {
-		if _, hasMaxTokens := mergedOptions["max_tokens"]; hasMaxTokens {
-			// Move max_tokens value to max_completion_tokens
-			mergedOptions["max_completion_tokens"] = mergedOptions["max_tokens"]
-			delete(mergedOptions, "max_tokens")
-		}
-	} else {
-		// For other models, ensure we use max_tokens and not max_completion_tokens
-		if _, hasMaxCompletionTokens := mergedOptions["max_completion_tokens"]; hasMaxCompletionTokens {
-			// Move max_completion_tokens value to max_tokens
-			mergedOptions["max_tokens"] = mergedOptions["max_completion_tokens"]
-			delete(mergedOptions, "max_completion_tokens")
-		}
-	}
-
-	// Add merged options to the request
-	for k, v := range mergedOptions {
-		request[k] = v
-	}
-
-	return json.Marshal(request)
-}
-
-// PrepareRequestWithSchema creates a request that includes JSON schema validation.
-// This uses OpenAI's function calling feature to enforce response structure.
-//
-// Parameters:
-//   - prompt: The input text or conversation
-//   - options: Additional request parameters
-//   - schema: JSON schema for response validation
-//
-// Returns:
-//   - Serialized JSON request body
-//   - Any error encountered during preparation
-func (p *OpenAIProvider) PrepareRequestWithSchema(prompt string, options map[string]interface{}, schema interface{}) ([]byte, error) {
-	p.logger.Debug("Preparing request with schema", "prompt", prompt, "schema", schema)
-
-	// First, ensure we have a proper object for the schema
-	var schemaObj interface{}
-	switch s := schema.(type) {
-	case string:
-		if err := json.Unmarshal([]byte(s), &schemaObj); err != nil {
-			return nil, fmt.Errorf("failed to unmarshal schema string: %w", err)
-		}
-	case []byte:
-		if err := json.Unmarshal(s, &schemaObj); err != nil {
-			return nil, fmt.Errorf("failed to unmarshal schema bytes: %w", err)
-		}
-	case map[string]interface{}:
-		schemaObj = s
-	default:
-		// Try to marshal and unmarshal to ensure we have a proper object
-		schemaBytes, err := json.Marshal(schema)
-		if err != nil {
-			return nil, fmt.Errorf("failed to marshal schema: %w", err)
-		}
-		if err := json.Unmarshal(schemaBytes, &schemaObj); err != nil {
-			return nil, fmt.Errorf("failed to unmarshal schema: %w", err)
-		}
-	}
-
-	// Clean the schema for OpenAI by removing unsupported validation rules
-	cleanSchema := cleanSchemaForOpenAI(schemaObj)
-
-	// Debug log the cleaned schema
-	cleanSchemaJSON, _ := json.MarshalIndent(cleanSchema, "", "  ")
-	p.logger.Debug("Cleaned schema for OpenAI", "schema", string(cleanSchemaJSON))
-
-	request := map[string]interface{}{
-		"model": p.model,
-		"messages": []map[string]interface{}{
-			{"role": "user", "content": prompt},
-		},
-		"response_format": map[string]interface{}{
-			"type": "json_schema",
-			"json_schema": map[string]interface{}{
-				"name":   "structured_response",
-				"schema": cleanSchema,
-				"strict": true,
-			},
-		},
-	}
-
-	// Handle system prompt as system message
-	if systemPrompt, ok := options["system_prompt"].(string); ok && systemPrompt != "" {
-		request["messages"] = append([]map[string]interface{}{
-			{"role": "system", "content": systemPrompt},
-		}, request["messages"].([]map[string]interface{})...)
-	}
-
-	// Create a merged copy of options to handle token parameters properly
-	mergedOptions := make(map[string]interface{})
-
-	// First add options from provider (p.options)
-	for k, v := range p.options {
-		if k != "system_prompt" {
-			mergedOptions[k] = v
-		}
-	}
-
-	// Then add options from the function parameters (may override provider options)
-	for k, v := range options {
-		if k != "system_prompt" {
-			mergedOptions[k] = v
-		}
-	}
-
-	// Handle max_tokens/max_completion_tokens conflict
-	// For models that need max_completion_tokens, ensure we use that and not max_tokens
-	if p.needsMaxCompletionTokens() {
-		if _, hasMaxTokens := mergedOptions["max_tokens"]; hasMaxTokens {
-			// Move max_tokens value to max_completion_tokens
-			mergedOptions["max_completion_tokens"] = mergedOptions["max_tokens"]
-			delete(mergedOptions, "max_tokens")
-		}
-	} else {
-		// For other models, ensure we use max_tokens and not max_completion_tokens
-		if _, hasMaxCompletionTokens := mergedOptions["max_completion_tokens"]; hasMaxCompletionTokens {
-			// Move max_completion_tokens value to max_tokens
-			mergedOptions["max_tokens"] = mergedOptions["max_completion_tokens"]
-			delete(mergedOptions, "max_completion_tokens")
-		}
-	}
-
-	// Add merged options to the request
-	for k, v := range mergedOptions {
-		request[k] = v
-	}
-
-	reqJSON, err := json.Marshal(request)
+	data, err := json.Marshal(requestBody)
 	if err != nil {
-		p.logger.Error("Failed to marshal request with schema", "error", err)
-		return nil, err
+		return nil, fmt.Errorf("failed to marshal request body: %w", err)
 	}
-
-	p.logger.Debug("Full request to OpenAI", "request", string(reqJSON))
-	return reqJSON, nil
-}
-
-// cleanSchemaForOpenAI removes validation rules that OpenAI doesn't support
-func cleanSchemaForOpenAI(schema interface{}) interface{} {
-	if schemaMap, ok := schema.(map[string]interface{}); ok {
-		result := make(map[string]interface{})
-		for k, v := range schemaMap {
-			switch k {
-			case "type", "properties", "required", "items":
-				if k == "properties" {
-					props := make(map[string]interface{})
-					if propsMap, ok := v.(map[string]interface{}); ok {
-						for propName, propSchema := range propsMap {
-							props[propName] = cleanSchemaForOpenAI(propSchema)
-						}
-					}
-					result[k] = props
-				} else if k == "items" {
-					result[k] = cleanSchemaForOpenAI(v)
-				} else {
-					result[k] = v
-				}
-			}
-		}
-		// Add additionalProperties: false at each object level
-		if schemaMap["type"] == "object" {
-			result["additionalProperties"] = false
-		}
-		return result
-	}
-	return schema
+	return data, nil
 }
 
 // ParseResponse extracts the generated text from the OpenAI API response.
 // It handles various response formats and error cases.
-//
-// Parameters:
-//   - body: Raw API response body
-//
-// Returns:
-//   - Generated text content
-//   - Any error encountered during parsing
-func (p *OpenAIProvider) ParseResponse(body []byte) (string, error) {
-	var response struct {
-		Choices []struct {
-			Message struct {
-				Content   string `json:"content"`
-				ToolCalls []struct {
-					ID       string `json:"id"`
-					Type     string `json:"type"`
-					Function struct {
-						Name      string          `json:"name"`
-						Arguments json.RawMessage `json:"arguments"`
-					} `json:"function"`
-				} `json:"tool_calls"`
-			} `json:"message"`
-		} `json:"choices"`
-	}
+func (p *OpenAIProvider) ParseResponse(body []byte) (*Response, error) {
+	response := openAIResponse{}
 
 	if err := json.Unmarshal(body, &response); err != nil {
-		return "", err
+		return nil, fmt.Errorf("failed to unmarshal response: %w", err)
 	}
 
 	if len(response.Choices) == 0 {
-		return "", fmt.Errorf("empty response from API")
+		return nil, errors.New("empty response from API")
+	}
+
+	usage := &Usage{}
+
+	if response.Usage != nil && response.Usage.PromptTokensDetails != nil {
+		usage = NewUsage(
+			response.Usage.PromptTokens,
+			response.Usage.PromptTokensDetails.CacheTokens,
+			response.Usage.CompletionTokens,
+			0,
+			0, // ReasoningTokens
+		)
 	}
 
 	message := response.Choices[0].Message
 	if message.Content != "" {
-		return message.Content, nil
+		return &Response{
+			Content: Text{message.Content},
+			Usage:   usage,
+		}, nil
 	}
 
 	if len(message.ToolCalls) > 0 {
 		var functionCalls []string
 		for _, call := range message.ToolCalls {
 			// Parse arguments as raw JSON to preserve the exact format
-			var args interface{}
+			var args any
 			if err := json.Unmarshal(call.Function.Arguments, &args); err != nil {
-				return "", fmt.Errorf("error parsing function arguments: %w", err)
+				return nil, fmt.Errorf("error parsing function arguments: %w", err)
 			}
 
-			functionCall, err := utils.FormatFunctionCall(call.Function.Name, args)
+			functionCall, err := FormatFunctionCall(call.Function.Name, args)
 			if err != nil {
-				return "", fmt.Errorf("error formatting function call: %w", err)
+				return nil, fmt.Errorf("error formatting function call: %w", err)
 			}
 			functionCalls = append(functionCalls, functionCall)
 		}
-		return strings.Join(functionCalls, "\n"), nil
+
+		return &Response{
+			Content: Text{strings.Join(functionCalls, "\n")},
+			Usage:   usage,
+		}, nil
 	}
 
-	return "", fmt.Errorf("no content or tool calls in response")
-}
-
-// HandleFunctionCalls processes function calling in the response.
-// This supports OpenAI's function calling and JSON mode features.
-func (p *OpenAIProvider) HandleFunctionCalls(body []byte) ([]byte, error) {
-	response := string(body)
-	functionCalls, err := utils.ExtractFunctionCalls(response)
-	if err != nil {
-		return nil, fmt.Errorf("error extracting function calls: %w", err)
-	}
-
-	if len(functionCalls) == 0 {
-		return nil, fmt.Errorf("no function calls found in response")
-	}
-
-	p.logger.Debug("Function calls to handle", "calls", functionCalls)
-	return json.Marshal(functionCalls)
+	return nil, errors.New("no content or tool calls in response")
 }
 
 // SetExtraHeaders configures additional HTTP headers for API requests.
@@ -473,210 +362,364 @@ func (p *OpenAIProvider) SetExtraHeaders(extraHeaders map[string]string) {
 	p.logger.Debug("Extra headers set", "headers", extraHeaders)
 }
 
-// SupportsStreaming indicates whether streaming is supported
-func (p *OpenAIProvider) SupportsStreaming() bool {
-	return true
-}
-
 // PrepareStreamRequest creates a request body for streaming API calls
-func (p *OpenAIProvider) PrepareStreamRequest(prompt string, options map[string]interface{}) ([]byte, error) {
-	// Start with regular request preparation
-	requestBody := map[string]interface{}{
-		"model": p.model,
-		"messages": []map[string]string{
-			{"role": "user", "content": prompt},
-		},
-		"stream": true,
+func (p *OpenAIProvider) PrepareStreamRequest(req *Request, options map[string]any) ([]byte, error) {
+	// Determine which model to use
+	model := p.model
+	if req.Model != "" {
+		model = req.Model
+	} else if m, ok := options["model"].(string); ok && m != "" {
+		model = m
 	}
 
-	// Create a merged copy of options to handle token parameters properly
-	mergedOptions := make(map[string]interface{})
-
-	// First add options from provider (p.options)
-	for k, v := range p.options {
-		if k != "stream" { // Don't override stream setting
-			mergedOptions[k] = v
-		}
+	if !p.HasCapability(CapStreaming, model) {
+		return nil, errors.New("streaming is not supported by this provider")
 	}
 
-	// Then add options from the function parameters (may override provider options)
-	for k, v := range options {
-		if k != "stream" { // Don't override stream setting
-			mergedOptions[k] = v
-		}
+	requestBody := p.initializeOpenAIRequestWithModel(model)
+	requestBody[openAIKeyStream] = true
+	requestBody["stream_options"] = map[string]bool{"include_usage": true}
+
+	// Handle system prompt from Request or options
+	systemPrompt := p.extractSystemPromptFromRequest(req, options)
+	if systemPrompt != "" {
+		p.addSystemPromptToRequestBody(requestBody, systemPrompt)
 	}
 
-	// Handle max_tokens/max_completion_tokens conflict
-	// For models that need max_completion_tokens, ensure we use that and not max_tokens
-	if p.needsMaxCompletionTokens() {
-		if _, hasMaxTokens := mergedOptions["max_tokens"]; hasMaxTokens {
-			// Move max_tokens value to max_completion_tokens
-			mergedOptions["max_completion_tokens"] = mergedOptions["max_tokens"]
-			delete(mergedOptions, "max_tokens")
-		}
-	} else {
-		// For other models, ensure we use max_tokens and not max_completion_tokens
-		if _, hasMaxCompletionTokens := mergedOptions["max_completion_tokens"]; hasMaxCompletionTokens {
-			// Move max_completion_tokens value to max_tokens
-			mergedOptions["max_tokens"] = mergedOptions["max_completion_tokens"]
-			delete(mergedOptions, "max_completion_tokens")
-		}
+	// Add messages from the Request
+	p.addMessagesToRequestBody(requestBody, req.Messages, options)
+
+	// Handle tools if present in options
+	p.handleToolsForRequest(requestBody, options)
+
+	// Handle structured response schema
+	if req.ResponseSchema != nil && p.HasCapability(CapStructuredResponse, model) {
+		p.addStructuredResponseToRequest(requestBody, req.ResponseSchema)
 	}
 
-	// Add merged options to the request
-	for k, v := range mergedOptions {
-		requestBody[k] = v
-	}
+	// Add remaining options
+	p.addRemainingOptions(requestBody, options)
 
-	return json.Marshal(requestBody)
+	data, err := json.Marshal(requestBody)
+	if err != nil {
+		return nil, fmt.Errorf("failed to marshal request body: %w", err)
+	}
+	return data, nil
 }
 
 // ParseStreamResponse processes a single chunk from a streaming response
-func (p *OpenAIProvider) ParseStreamResponse(chunk []byte) (string, error) {
+func (p *OpenAIProvider) ParseStreamResponse(chunk []byte) (*Response, error) {
 	// Skip empty lines
 	if len(bytes.TrimSpace(chunk)) == 0 {
-		return "", fmt.Errorf("empty chunk")
+		return nil, errors.New("empty chunk")
 	}
 
 	// Check for [DONE] marker
 	if bytes.Equal(bytes.TrimSpace(chunk), []byte("[DONE]")) {
-		return "", io.EOF
+		return nil, io.EOF
 	}
 
 	// Parse the chunk
-	var response struct {
-		Choices []struct {
-			Delta struct {
-				Role    string `json:"role"`
-				Content string `json:"content"`
-			} `json:"delta"`
-			FinishReason string `json:"finish_reason"`
-		} `json:"choices"`
-	}
+	response := openAIStreamResponse{}
 
 	if err := json.Unmarshal(chunk, &response); err != nil {
-		return "", fmt.Errorf("malformed response: %w", err)
+		return nil, fmt.Errorf("malformed response: %w", err)
 	}
 
 	if len(response.Choices) == 0 {
-		return "", fmt.Errorf("no choices in response")
+		return nil, errors.New("no choices in response")
 	}
 
 	// Handle finish reason
 	if response.Choices[0].FinishReason != "" {
-		return "", io.EOF
+		return nil, io.EOF
 	}
 
 	// Skip role-only messages
 	if response.Choices[0].Delta.Role != "" && response.Choices[0].Delta.Content == "" {
-		return "", fmt.Errorf("skip token")
+		return nil, errors.New("skip token")
 	}
 
-	return response.Choices[0].Delta.Content, nil
+	usage := &Usage{}
+
+	if response.Usage != nil && response.Usage.PromptTokensDetails != nil {
+		usage = NewUsage(
+			response.Usage.PromptTokens,
+			response.Usage.PromptTokensDetails.CacheTokens,
+			response.Usage.CompletionTokens,
+			0,
+			0, // ReasoningTokens
+		)
+	}
+
+	return &Response{
+		Content: Text{
+			response.Choices[0].Delta.Content,
+		},
+		Usage: usage,
+	}, nil
 }
 
-// PrepareRequestWithMessages creates a request body using structured message objects
-// rather than a flattened prompt string. This enables more efficient caching and
-// better preserves conversation structure for the OpenAI API.
-//
-// Parameters:
-//   - messages: Slice of MemoryMessage objects representing the conversation
-//   - options: Additional options for the request
-//
-// Returns:
-//   - Serialized JSON request body
-//   - Any error encountered during preparation
-func (p *OpenAIProvider) PrepareRequestWithMessages(messages []types.MemoryMessage, options map[string]interface{}) ([]byte, error) {
-	request := map[string]interface{}{
-		"model":    p.model,
-		"messages": []map[string]interface{}{},
+// needsMaxCompletionTokens checks if the model requires max_completion_tokens instead of max_tokens
+func (p *OpenAIProvider) needsMaxCompletionTokens() bool {
+	if strings.HasPrefix(p.model, "o") {
+		return true
 	}
 
-	// Handle system prompt as system message
-	if systemPrompt, ok := options["system_prompt"].(string); ok && systemPrompt != "" {
-		request["messages"] = append(request["messages"].([]map[string]interface{}), map[string]interface{}{
+	if strings.Contains(p.model, "4o") || strings.Contains(p.model, "-o") {
+		return true
+	}
+
+	return false
+}
+
+// initializeOpenAIRequestWithModel creates the base request structure with specified model
+func (p *OpenAIProvider) initializeOpenAIRequestWithModel(model string) map[string]any {
+	return map[string]any{
+		"model":    model,
+		"messages": []map[string]any{},
+	}
+}
+
+// extractSystemPromptFromRequest gets system prompt from request or options
+func (p *OpenAIProvider) extractSystemPromptFromRequest(req *Request, options map[string]any) string {
+	// Priority: Request.SystemPrompt > options["system_prompt"]
+	if req.SystemPrompt != "" {
+		return req.SystemPrompt
+	}
+	if sp, ok := options[openAIKeySystemPrompt].(string); ok && sp != "" {
+		return sp
+	}
+	return ""
+}
+
+// addSystemPromptToRequestBody adds the system prompt to the request
+func (p *OpenAIProvider) addSystemPromptToRequestBody(requestBody map[string]any, systemPrompt string) {
+	if systemPrompt == "" {
+		return
+	}
+
+	if messages, ok := requestBody["messages"].([]map[string]any); ok {
+		systemMessage := map[string]any{
 			"role":    "system",
 			"content": systemPrompt,
-		})
-	}
-
-	// Convert MemoryMessage objects to OpenAI messages format
-	for _, msg := range messages {
-		message := map[string]interface{}{
-			"role":    msg.Role,
-			"content": msg.Content,
 		}
+		requestBody["messages"] = append([]map[string]any{systemMessage}, messages...)
+	}
+}
 
-		// Add metadata if present
-		if len(msg.Metadata) > 0 {
-			for k, v := range msg.Metadata {
-				message[k] = v
-			}
-		}
-
-		request["messages"] = append(request["messages"].([]map[string]interface{}), message)
+// addMessagesToRequestBody converts and adds messages to the request
+func (p *OpenAIProvider) addMessagesToRequestBody(
+	requestBody map[string]any,
+	messages []Message,
+	options map[string]any,
+) {
+	if len(messages) == 0 {
+		return
 	}
 
-	// Handle tool_choice
-	if toolChoice, ok := options["tool_choice"].(string); ok {
-		request["tool_choice"] = toolChoice
+	openAIMessages := make([]map[string]any, 0, len(messages))
+	for i := range messages {
+		openAIMsg := p.convertMessageToOpenAIFormat(&messages[i], options)
+		openAIMessages = append(openAIMessages, openAIMsg)
 	}
 
-	// Handle tools
-	if tools, ok := options["tools"].([]utils.Tool); ok && len(tools) > 0 {
-		openAITools := make([]map[string]interface{}, len(tools))
-		for i, tool := range tools {
-			openAITools[i] = map[string]interface{}{
-				"type": "function",
-				"function": map[string]interface{}{
-					"name":        tool.Function.Name,
-					"description": tool.Function.Description,
-					"parameters":  tool.Function.Parameters,
+	if existingMessages, ok := requestBody["messages"].([]map[string]any); ok {
+		requestBody["messages"] = append(existingMessages, openAIMessages...)
+	} else {
+		requestBody["messages"] = openAIMessages
+	}
+}
+
+// convertMessageToOpenAIFormat converts a Message to OpenAI's format
+func (p *OpenAIProvider) convertMessageToOpenAIFormat(msg *Message, _ map[string]any) map[string]any {
+	openAIMsg := map[string]any{
+		"role":    msg.Role,
+		"content": msg.Content,
+	}
+
+	// Add name if present
+	if msg.Name != "" {
+		openAIMsg["name"] = msg.Name
+	}
+
+	// Add tool_call_id if present
+	if msg.ToolCallID != "" {
+		openAIMsg["tool_call_id"] = msg.ToolCallID
+	}
+
+	// Handle tool calls if present
+	if len(msg.ToolCalls) > 0 {
+		openAIToolCalls := make([]map[string]any, len(msg.ToolCalls))
+		for i, toolCall := range msg.ToolCalls {
+			openAIToolCalls[i] = map[string]any{
+				"id":   toolCall.ID,
+				"type": toolCall.Type,
+				"function": map[string]any{
+					"name":      toolCall.Function.Name,
+					"arguments": string(toolCall.Function.Arguments),
 				},
-				"strict": true, // Add this if you want strict mode
 			}
 		}
-		request["tools"] = openAITools
+		openAIMsg["tool_calls"] = openAIToolCalls
 	}
 
-	// Create a merged copy of options to handle token parameters properly
-	mergedOptions := make(map[string]interface{})
+	return openAIMsg
+}
 
-	// First add options from provider (p.options)
+// handleToolsForRequest processes tools and adds them to the request
+func (p *OpenAIProvider) handleToolsForRequest(requestBody map[string]any, options map[string]any) {
+	tools, ok := options[openAIKeyTools].([]models.Tool)
+	if !ok || len(tools) == 0 {
+		return
+	}
+
+	// Convert tools to OpenAI format
+	openAITools := make([]map[string]any, len(tools))
+	for i, tool := range tools {
+		openAITools[i] = map[string]any{
+			"type": "function",
+			"function": map[string]any{
+				"name":        tool.Function.Name,
+				"description": tool.Function.Description,
+				"parameters":  tool.Function.Parameters,
+			},
+			"strict": true,
+		}
+	}
+
+	requestBody["tools"] = openAITools
+
+	// Handle tool_choice if specified
+	if toolChoice, ok := options[openAIKeyToolChoice].(string); ok {
+		requestBody["tool_choice"] = toolChoice
+	}
+}
+
+// addStructuredResponseToRequest adds structured response schema to the request
+func (p *OpenAIProvider) addStructuredResponseToRequest(requestBody map[string]any, schema any) {
+	// For OpenAI, we use response_format with JSON schema
+	requestBody["response_format"] = map[string]any{
+		"type": "json_schema",
+		"json_schema": map[string]any{
+			"name":   "response",
+			"schema": schema,
+			"strict": true,
+		},
+	}
+}
+
+// addRemainingOptions adds any remaining options to the request body
+func (p *OpenAIProvider) addRemainingOptions(requestBody map[string]any, options map[string]any) {
+	// Create merged options excluding handled keys
+	mergedOptions := make(map[string]any)
+
+	// Add provider options first
 	for k, v := range p.options {
-		if k != "tools" && k != "tool_choice" && k != "system_prompt" && k != "structured_messages" {
+		if !p.isGlobalOption(k) {
 			mergedOptions[k] = v
 		}
 	}
 
-	// Then add options from the function parameters (may override provider options)
+	// Add request options (may override provider options)
 	for k, v := range options {
-		if k != "tools" && k != "tool_choice" && k != "system_prompt" && k != "structured_messages" {
+		if !p.isGlobalOption(k) {
 			mergedOptions[k] = v
 		}
 	}
 
-	// Handle max_tokens/max_completion_tokens conflict
+	// Handle token parameters
+	p.handleTokenParameters(mergedOptions)
+
+	// Add all merged options to request body
+	for k, v := range mergedOptions {
+		requestBody[k] = v
+	}
+}
+
+// isGlobalOption checks if a key is handled globally and should be excluded from options
+func (p *OpenAIProvider) isGlobalOption(key string) bool {
+	return key == openAIKeySystemPrompt ||
+		key == openAIKeyTools ||
+		key == openAIKeyToolChoice ||
+		key == openAIKeyStructuredMessages ||
+		key == openAIKeyStream
+}
+
+// handleTokenParameters handles max_tokens/max_completion_tokens conflict
+func (p *OpenAIProvider) handleTokenParameters(mergedOptions map[string]any) {
 	// For models that need max_completion_tokens, ensure we use that and not max_tokens
 	if p.needsMaxCompletionTokens() {
-		if _, hasMaxTokens := mergedOptions["max_tokens"]; hasMaxTokens {
+		if _, hasMaxTokens := mergedOptions[openAIKeyMaxTokens]; hasMaxTokens {
 			// Move max_tokens value to max_completion_tokens
-			mergedOptions["max_completion_tokens"] = mergedOptions["max_tokens"]
-			delete(mergedOptions, "max_tokens")
+			mergedOptions["max_completion_tokens"] = mergedOptions[openAIKeyMaxTokens]
+			delete(mergedOptions, openAIKeyMaxTokens)
 		}
 	} else {
 		// For other models, ensure we use max_tokens and not max_completion_tokens
 		if _, hasMaxCompletionTokens := mergedOptions["max_completion_tokens"]; hasMaxCompletionTokens {
 			// Move max_completion_tokens value to max_tokens
-			mergedOptions["max_tokens"] = mergedOptions["max_completion_tokens"]
+			mergedOptions[openAIKeyMaxTokens] = mergedOptions["max_completion_tokens"]
 			delete(mergedOptions, "max_completion_tokens")
 		}
 	}
+}
 
-	// Add merged options to the request
-	for k, v := range mergedOptions {
-		request[k] = v
-	}
+type openAIResponse struct {
+	Usage   *openAIUsage   `json:"usage"`
+	Choices []openAIChoice `json:"choices"`
+}
 
-	return json.Marshal(request)
+type openAIChoice struct {
+	Message *openAIMessage `json:"message"`
+}
+
+type openAIMessage struct {
+	Content   string           `json:"content"`
+	ToolCalls []openAIToolCall `json:"tool_calls"`
+}
+
+type openAIToolCall struct {
+	Function *openAIFunction `json:"function"`
+	ID       string          `json:"id"`
+	Type     string          `json:"type"`
+}
+
+type openAIFunction struct {
+	Name      string          `json:"name"`
+	Arguments json.RawMessage `json:"arguments"`
+}
+type openAIUsage struct {
+	PromptTokensDetails     *openAIPromptTokensDetails     `json:"prompt_tokens_details"`
+	CompletionTokensDetails *openAICompletionTokensDetails `json:"completion_tokens_details"`
+	PromptTokens            int64                          `json:"prompt_tokens"`
+	CompletionTokens        int64                          `json:"completion_tokens"`
+	TotalTokens             int64                          `json:"total_tokens"`
+}
+
+type openAIPromptTokensDetails struct {
+	CacheTokens int64 `json:"cache_tokens"`
+	AudioTokens int64 `json:"audio_tokens"`
+}
+
+type openAICompletionTokensDetails struct {
+	ReasoningTokens          int64 `json:"reasoning_tokens"`
+	AudioTokens              int64 `json:"audio_tokens"`
+	AcceptedPredictionTokens int64 `json:"accepted_prediction_tokens"`
+	RejectedPredictionTokens int64 `json:"rejected_prediction_tokens"`
+}
+
+type openAIStreamResponse struct {
+	Usage   *openAIUsage         `json:"usage,omitempty"`
+	Choices []openAIStreamChoice `json:"choices"`
+}
+
+type openAIStreamChoice struct {
+	Delta        openAIStreamDelta `json:"delta"`
+	FinishReason string            `json:"finish_reason"`
+}
+
+type openAIStreamDelta struct {
+	Role    string `json:"role"`
+	Content string `json:"content"`
 }
