@@ -5,33 +5,32 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/teilomillet/gollm/types"
 )
 
 // TestAllProvidersImplementStructuredMessages verifies that all providers
 // properly implement the PrepareRequestWithMessages method
 func TestAllProvidersImplementStructuredMessages(t *testing.T) {
 	// Define test messages
-	testMessages := []types.MemoryMessage{
+	testMessages := []Message{
 		{
-			Role:         "user",
-			Content:      "Hello, how are you?",
-			CacheControl: "ephemeral",
+			Role:      "user",
+			Content:   "Hello, how are you?",
+			CacheType: "ephemeral",
 		},
 		{
-			Role:         "assistant",
-			Content:      "I'm doing well, thank you for asking!",
-			CacheControl: "",
+			Role:      "assistant",
+			Content:   "I'm doing well, thank you for asking!",
+			CacheType: "",
 		},
 		{
-			Role:         "user",
-			Content:      "Tell me about structured messages",
-			CacheControl: "ephemeral",
+			Role:      "user",
+			Content:   "Tell me about structured messages",
+			CacheType: "ephemeral",
 		},
 	}
 
 	// Test options
-	options := map[string]interface{}{
+	options := map[string]any{
 		"system_prompt": "You are a helpful assistant",
 		"temperature":   0.7,
 		"max_tokens":    1024,
@@ -52,19 +51,17 @@ func TestAllProvidersImplementStructuredMessages(t *testing.T) {
 
 	for _, p := range providers {
 		t.Run(p.name, func(t *testing.T) {
+			builder := NewRequestBuilder()
+			builder.WithMessages(testMessages)
+			request := builder.Build()
+
 			// Make sure the PrepareRequestWithMessages method produces valid output
-			result, err := p.provider.PrepareRequestWithMessages(testMessages, options)
+			result, err := p.provider.PrepareRequest(request, options)
 			require.NoError(t, err, "PrepareRequestWithMessages should not fail")
 			require.NotNil(t, result, "PrepareRequestWithMessages should return non-nil result")
 
 			// Result should be valid JSON
-			assert.True(t, len(result) > 0, "Result should not be empty")
-
-			// Test the type assertion works (the provider correctly implements the method)
-			_, ok := p.provider.(interface {
-				PrepareRequestWithMessages(messages []types.MemoryMessage, options map[string]interface{}) ([]byte, error)
-			})
-			assert.True(t, ok, "Provider should implement PrepareRequestWithMessages")
+			assert.NotEmpty(t, result, "Result should not be empty")
 		})
 	}
 }
@@ -75,10 +72,14 @@ func TestStructuredMessagesFormat(t *testing.T) {
 	// Test with minimal providers to verify message format
 	t.Run("OpenAI format", func(t *testing.T) {
 		provider := NewOpenAIProvider("fake-key", "gpt-4", nil)
-		messages := []types.MemoryMessage{
+		messages := []Message{
 			{Role: "user", Content: "Hello"},
 		}
-		result, err := provider.PrepareRequestWithMessages(messages, nil)
+		builder := NewRequestBuilder()
+		builder.WithMessages(messages)
+		request := builder.Build()
+
+		result, err := provider.PrepareRequest(request, nil)
 		require.NoError(t, err)
 
 		// Should contain the message format in OpenAI style
@@ -88,10 +89,15 @@ func TestStructuredMessagesFormat(t *testing.T) {
 
 	t.Run("Anthropic format", func(t *testing.T) {
 		provider := NewAnthropicProvider("fake-key", "claude-3", nil)
-		messages := []types.MemoryMessage{
-			{Role: "user", Content: "Hello", CacheControl: "ephemeral"},
+		messages := []Message{
+			{Role: "user", Content: "Hello", CacheType: "ephemeral"},
 		}
-		result, err := provider.PrepareRequestWithMessages(messages, nil)
+
+		builder := NewRequestBuilder()
+		builder.WithMessages(messages)
+		request := builder.Build()
+
+		result, err := provider.PrepareRequest(request, nil)
 		require.NoError(t, err)
 
 		// Should contain cache_control for Anthropic
