@@ -4,7 +4,9 @@
 package gollm
 
 import (
-	"github.com/teilomillet/gollm/llm"
+	"fmt"
+
+	"github.com/weave-labs/gollm/llm"
 )
 
 // Validate checks if the given struct is valid according to its validation rules.
@@ -18,7 +20,7 @@ import (
 //   - email: Email format validation
 //   - url: URL format validation
 //   - regex: Pattern matching
-//   - contains/excludes: String content validation
+//   - contains/excludes: AsText content validation
 //   - unique: Array unique items validation
 //   - minItems/maxItems: Array length validation
 //   - password: Password strength validation
@@ -31,7 +33,7 @@ import (
 //	    Email     string `validate:"required,email"`
 //	    Password  string `validate:"required,password=strong"`
 //	}
-//	
+//
 //	config := Config{
 //	    Model:     "gpt-4",
 //	    MaxTokens: 2048,
@@ -45,8 +47,11 @@ import (
 //
 // Returns:
 //   - error: nil if validation passes, otherwise returns detailed validation errors
-func Validate(s interface{}) error {
-	return llm.Validate(s)
+func Validate(s any) error {
+	if err := llm.Validate(s); err != nil {
+		return fmt.Errorf("validation failed: %w", err)
+	}
+	return nil
 }
 
 // GenerateJSONSchema generates a JSON schema for the given struct.
@@ -70,13 +75,13 @@ func Validate(s interface{}) error {
 //	    Tokens   int      `json:"tokens,omitempty" validate:"min=0"`
 //	    Tags     []string `json:"tags,omitempty" validate:"unique"`
 //	}
-//	
+//
 //	type Conversation struct {
 //	    ID       string    `json:"id" validate:"required,uuid"`
 //	    Messages []Message `json:"messages" validate:"required,min=1"`
 //	    Model    string    `json:"model" validate:"required,model"`
 //	}
-//	
+//
 //	schema, err := GenerateJSONSchema(&Conversation{})
 //
 // Parameters:
@@ -85,6 +90,10 @@ func Validate(s interface{}) error {
 // Returns:
 //   - []byte: The generated JSON schema as a byte slice
 //   - error: Any error encountered during schema generation
-func GenerateJSONSchema(v interface{}) ([]byte, error) {
-	return llm.GenerateJSONSchema(v)
+func GenerateJSONSchema(v any) ([]byte, error) {
+	schema, err := llm.GenerateJSONSchema(v)
+	if err != nil {
+		return nil, fmt.Errorf("failed to generate JSON schema: %w", err)
+	}
+	return schema, nil
 }
