@@ -9,7 +9,8 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/teilomillet/gollm"
+
+	"github.com/weave-labs/gollm"
 )
 
 // TestProviderIntegration demonstrates comprehensive testing across providers
@@ -73,36 +74,30 @@ func TestJSONValidation(t *testing.T) {
 			BatchTimeout: 5 * time.Minute,
 		})
 
-	// Generate schema from struct
-	schema, err := gollm.GenerateJSONSchema(SOLIDResponse{})
-	if err != nil {
-		t.Fatalf("Failed to generate JSON schema: %v", err)
-	}
-
 	// Test JSON response with schema validation
-	test.AddCase("json_solid_principles",
-		fmt.Sprintf(`Please provide the SOLID principles and examples.
-		Format your response as a JSON object that adheres to this schema:
-		%s
+	tc := SetExpectedSchema[SOLIDResponse](test.AddCase("json_solid_principles",
+		`Please provide the SOLID principles and examples.
+		Format your response as a JSON object with:
+		- principles: array of strings, each containing one SOLID principle
+		- examples: array of strings, each containing a practical example
 		
 		The principles array should contain each SOLID principle as a simple string.
 		The examples array should contain practical examples as simple one-line strings.
-		Do not include objects or complex types in the arrays.`, string(schema))).
+		Do not include objects or complex types in the arrays.`).
 		WithTimeout(60*time.Second).
-		WithOption("max_tokens", 2000).
-		ExpectSchema(schema).
-		Validate(func(response string) error {
-			var data SOLIDResponse
-			if err := json.Unmarshal([]byte(response), &data); err != nil {
-				return fmt.Errorf("invalid JSON response: %v", err)
-			}
+		WithOption("max_tokens", 2000))
+	tc.Validate(func(response string) error {
+		var data SOLIDResponse
+		if err := json.Unmarshal([]byte(response), &data); err != nil {
+			return fmt.Errorf("invalid JSON response: %w", err)
+		}
 
-			if err := gollm.Validate(&data); err != nil {
-				return fmt.Errorf("validation failed: %v", err)
-			}
+		if err := gollm.Validate(&data); err != nil {
+			return fmt.Errorf("validation failed: %w", err)
+		}
 
-			return nil
-		})
+		return nil
+	})
 
 	ctx := context.Background()
 	test.RunBatch(ctx)
@@ -142,24 +137,31 @@ func TestProviderSpecificFeatures(t *testing.T) {
 }
 
 // Add batch integration test
-func TestBatchIntegration(t *testing.T) {
-	test := NewTest(t).
-		WithProviders(map[string]string{
-			"anthropic": "claude-3-5-haiku-latest",
-			"openai":    "gpt-4o-mini",
-		}).
-		WithBatchConfig(BatchTestConfig{
-			EnableBatch:  true,
-			MaxParallel:  2,
-			BatchTimeout: 2 * time.Minute,
-		})
+// validatePrimeNumbers validates that at least 3 prime numbers are in the response
+func validatePrimeNumbers(response string) error {
+	// Define prime numbers up to 20
+	primes := []string{"2", "3", "5", "7", "11", "13", "17", "19"}
+	count := 0
 
-	// Test cases with different validation requirements
+	// Count how many prime numbers are in the response
+	for _, prime := range primes {
+		if strings.Contains(response, prime) {
+			count++
+		}
+	}
+
+	if count < 3 {
+		return fmt.Errorf("expected at least 3 prime numbers, found %d", count)
+	}
+	return nil
+}
+
+// setupBatchTestCases adds standard test cases to the test runner
+func setupBatchTestCases(test *TestRunner) {
 	testCases := []struct {
-		name      string
-		query     string
-		validate  ValidationFunc
-		expectErr bool
+		name     string
+		query    string
+		validate ValidationFunc
 	}{
 		{
 			name:     "simple_math",
@@ -172,25 +174,9 @@ func TestBatchIntegration(t *testing.T) {
 			validate: ExpectContains("def fibonacci"),
 		},
 		{
-			name:  "pattern_matching",
-			query: "List three prime numbers between 1 and 20",
-			validate: func(response string) error {
-				// Define prime numbers up to 20
-				primes := []string{"2", "3", "5", "7", "11", "13", "17", "19"}
-				count := 0
-
-				// Count how many prime numbers are in the response
-				for _, prime := range primes {
-					if strings.Contains(response, prime) {
-						count++
-					}
-				}
-
-				if count < 3 {
-					return fmt.Errorf("expected at least 3 prime numbers, found %d", count)
-				}
-				return nil
-			},
+			name:     "pattern_matching",
+			query:    "List three prime numbers between 1 and 20",
+			validate: validatePrimeNumbers,
 		},
 	}
 
@@ -200,20 +186,19 @@ func TestBatchIntegration(t *testing.T) {
 			WithTimeout(30 * time.Second).
 			Validate(tc.validate)
 	}
+}
 
-	ctx := context.Background()
-	test.RunBatch(ctx)
-
-	metrics := test.GetBatchMetrics()
-
+// verifyBatchMetrics checks that batch execution metrics are valid
+func verifyBatchMetrics(t *testing.T, metrics *BatchMetrics) {
+	t.Helper()
 	// Verify batch execution completed
-	assert.True(t, metrics.BatchTiming.TotalDuration > 0)
+	assert.Positive(t, metrics.BatchTiming.TotalDuration)
 	assert.True(t, metrics.BatchTiming.EndTime.After(metrics.BatchTiming.StartTime))
 
 	// Check provider latencies
 	for provider, latency := range metrics.BatchTiming.ProviderLatency {
 		t.Logf("Provider %s average latency: %v", provider, latency)
-		assert.True(t, latency > 0)
+		assert.Positive(t, latency)
 	}
 
 	// Verify error handling
@@ -222,6 +207,27 @@ func TestBatchIntegration(t *testing.T) {
 			t.Logf("Provider %s error: %v", provider, err)
 		}
 	}
+}
+
+func TestBatchIntegration(t *testing.T) {
+	test := NewTest(t).
+		WithProviders(map[string]string{
+			"anthropic": "claude-3-5-haiku-latest",
+			"openai":    "gpt-4o-mini",
+		}).
+		WithBatchConfig(BatchTestConfig{
+			EnableBatch:  true,
+			MaxParallel:  2,
+			BatchTimeout: 2 * time.Minute,
+		})
+
+	setupBatchTestCases(test)
+
+	ctx := context.Background()
+	test.RunBatch(ctx)
+
+	metrics := test.GetBatchMetrics()
+	verifyBatchMetrics(t, metrics)
 }
 
 // TestBatchCrossProvider tests cross-provider consistency
