@@ -1,11 +1,13 @@
 // Package gollm provides a high-level interface for interacting with various Language Learning Models (LLMs).
 // This file re-exports configuration types and functions from the config package to provide
 // a clean, centralized API for configuring LLM interactions.
+//
+//nolint:gochecknoglobals // These are re-exports to simplify the API surface.
 package gollm
 
 import (
-	"github.com/teilomillet/gollm/config"
-	"github.com/teilomillet/gollm/utils"
+	"github.com/weave-labs/gollm/config"
+	"github.com/weave-labs/gollm/internal/logging"
 )
 
 // Re-export core configuration types for easier access
@@ -34,7 +36,7 @@ type (
 	// Example usage:
 	//   cfg := NewConfig()
 	//   cfg = ApplyOptions(cfg, SetLogLevel(LogLevelInfo))
-	LogLevel = utils.LogLevel
+	LogLevel = logging.LogLevel
 
 	// MemoryOption configures the memory settings for conversation history.
 	// It controls how much context is retained between interactions.
@@ -78,12 +80,13 @@ var (
 	SetAPIKey         = config.SetAPIKey         // Sets the API key for the current provider
 
 	// Generation parameters
-	SetTemperature      = config.SetTemperature      // Controls randomness in generation (0.0-1.0)
-	SetMaxTokens        = config.SetMaxTokens        // Sets maximum tokens to generate
-	SetTopP             = config.SetTopP             // Controls nucleus sampling
-	SetFrequencyPenalty = config.SetFrequencyPenalty // Penalizes frequent token usage
-	SetPresencePenalty  = config.SetPresencePenalty  // Penalizes repeated tokens
-	SetSeed             = config.SetSeed             // Sets random seed for reproducible generation
+	SetTemperature         = config.SetTemperature         // Controls randomness in generation (0.0-1.0)
+	SetMaxTokens           = config.SetMaxTokens           // Sets maximum tokens to generate
+	SetMaxCompletionTokens = config.SetMaxCompletionTokens // Sets maximum output tokens for generation
+	SetTopP                = config.SetTopP                // Controls nucleus sampling
+	SetFrequencyPenalty    = config.SetFrequencyPenalty    // Penalizes frequent token usage
+	SetPresencePenalty     = config.SetPresencePenalty     // Penalizes repeated tokens
+	SetSeed                = config.SetSeed                // Sets random seed for reproducible generation
 
 	// Advanced generation parameters
 	SetMinP          = config.SetMinP          // Sets minimum probability threshold
@@ -111,9 +114,8 @@ var (
 
 // LogLevel constants define available logging verbosity levels
 const (
-	LogLevelOff   = utils.LogLevelOff   // Disables all logging
-	LogLevelError = utils.LogLevelError // Logs only errors
-	LogLevelWarn  = utils.LogLevelWarn  // Logs warnings and errors
-	LogLevelInfo  = utils.LogLevelInfo  // Logs info, warnings, and errors
-	LogLevelDebug = utils.LogLevelDebug // Logs all messages including debug
+	LogLevelError = logging.LogLevelError // Logs only errors
+	LogLevelWarn  = logging.LogLevelWarn  // Logs warnings and errors
+	LogLevelInfo  = logging.LogLevelInfo  // Logs info, warnings, and errors
+	LogLevelDebug = logging.LogLevelDebug // Logs all messages including debug
 )
