@@ -6,7 +6,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/teilomillet/gollm/llm"
+	"github.com/weave-labs/gollm/llm"
 )
 
 // assessPrompt evaluates a prompt's quality and effectiveness using the configured LLM.
@@ -83,11 +83,11 @@ func (po *PromptOptimizer) assessPrompt(ctx context.Context, prompt *llm.Prompt)
 	}
 
 	// Parse and validate assessment response
-	cleanedResponse := cleanJSONResponse(response)
+	cleanedResponse := cleanJSONResponse(response.AsText())
 	var assessment PromptAssessment
 	err = json.Unmarshal([]byte(cleanedResponse), &assessment)
 	if err != nil {
-		po.debugManager.LogResponse(fmt.Sprintf("Raw response: %s", response))
+		po.debugManager.LogResponse("assessment_parse_error", "Raw response: "+response.AsText())
 		return OptimizationEntry{}, fmt.Errorf("failed to parse assessment response: %w", err)
 	}
 
@@ -128,27 +128,27 @@ func (po *PromptOptimizer) assessPrompt(ctx context.Context, prompt *llm.Prompt)
 // Example threshold values:
 //   - Numerical: 0.75 requires score >= 15/20
 //   - Letter: Requires A- or better
-func (po *PromptOptimizer) isOptimizationGoalMet(assessment PromptAssessment) (bool, error) {
+func (po *PromptOptimizer) isOptimizationGoalMet(assessment *PromptAssessment) (bool, error) {
 	if po.ratingSystem == "" {
 		return false, nil
 	}
 
 	switch po.ratingSystem {
 	case "numerical":
-		return assessment.OverallScore >= 20*po.threshold, nil
+		return assessment.OverallScore >= MaxRatingScale*po.threshold, nil
 	case "letter":
 		gradeValues := map[string]float64{
-			"A+": 4.3, "A": 4.0, "A-": 3.7,
-			"B+": 3.3, "B": 3.0, "B-": 2.7,
-			"C+": 2.3, "C": 2.0, "C-": 1.7,
-			"D+": 1.3, "D": 1.0, "D-": 0.7,
-			"F": 0.0,
+			"A+": GradeValueAPlus, "A": GradeValueA, "A-": GradeValueAMinus,
+			"B+": GradeValueBPlus, "B": GradeValueB, "B-": GradeValueBMinus,
+			"C+": GradeValueCPlus, "C": GradeValueC, "C-": GradeValueCMinus,
+			"D+": GradeValueDPlus, "D": GradeValueD, "D-": GradeValueDMinus,
+			"F": GradeValueF,
 		}
 		gradeValue, exists := gradeValues[assessment.OverallGrade]
 		if !exists {
 			return false, fmt.Errorf("invalid grade: %s", assessment.OverallGrade)
 		}
-		return gradeValue >= 3.7, nil // Equivalent to A- or better
+		return gradeValue >= MinimumOptimizationGradeValue, nil // Equivalent to A- or better
 	default:
 		return false, fmt.Errorf("unknown rating system: %s", po.ratingSystem)
 	}
