@@ -6,24 +6,17 @@ package optimizer
 import (
 	"time"
 
-	"github.com/teilomillet/gollm/llm"
-	"github.com/teilomillet/gollm/utils"
+	"github.com/weave-labs/gollm/internal/debug"
+	"github.com/weave-labs/gollm/llm"
 )
 
 // Metric represents a quantitative or qualitative measure of prompt performance.
 // Each metric provides a specific aspect of evaluation with supporting reasoning.
 type Metric struct {
-	// Name identifies the metric (e.g., "Clarity", "Specificity")
-	Name string `json:"name"`
-
-	// Description explains what the metric measures and its significance
-	Description string `json:"description"`
-
-	// Value is the numerical score (0-20 scale) assigned to this metric
-	Value float64 `json:"value" validate:"required,min=0,max=20"`
-
-	// Reasoning provides the rationale behind the assigned value
-	Reasoning string `json:"reasoning"`
+	Name        string  `json:"name"`
+	Description string  `json:"description"`
+	Reasoning   string  `json:"reasoning"`
+	Value       float64 `json:"value" validate:"required,min=0,max=20"`
 }
 
 // Strength represents a positive aspect of the prompt with a concrete example.
@@ -46,42 +39,22 @@ type Weakness struct {
 
 // Suggestion represents a proposed improvement to the prompt with impact estimation.
 type Suggestion struct {
-	// Description outlines the suggested change
-	Description string `json:"description"`
-
-	// ExpectedImpact estimates the improvement's effect (0-20 scale)
-	ExpectedImpact float64 `json:"expectedImpact" validate:"required,min=0,max=20"`
-
-	// Reasoning explains why this suggestion would improve the prompt
-	Reasoning string `json:"reasoning"`
+	Description    string  `json:"description"`
+	Reasoning      string  `json:"reasoning"`
+	ExpectedImpact float64 `json:"expected_impact" validate:"required,min=0,max=20"`
 }
 
 // PromptAssessment provides a comprehensive evaluation of a prompt's quality
 // including metrics, strengths, weaknesses, and suggestions for improvement.
 type PromptAssessment struct {
-	// Metrics contains specific performance measurements
-	Metrics []Metric `json:"metrics" validate:"required,min=1"`
-
-	// Strengths lists positive aspects of the prompt
-	Strengths []Strength `json:"strengths" validate:"required,min=1"`
-
-	// Weaknesses identifies areas needing improvement
-	Weaknesses []Weakness `json:"weaknesses" validate:"required,min=1"`
-
-	// Suggestions provides actionable improvements
-	Suggestions []Suggestion `json:"suggestions" validate:"required,min=1"`
-
-	// OverallScore represents the prompt's overall quality (0-20 scale)
-	OverallScore float64 `json:"overallScore" validate:"required,min=0,max=20"`
-
-	// OverallGrade provides a letter grade assessment (e.g., "A", "B+")
-	OverallGrade string `json:"overallGrade" validate:"required,validGrade"`
-
-	// EfficiencyScore measures token usage and processing efficiency
-	EfficiencyScore float64 `json:"efficiencyScore" validate:"required,min=0,max=20"`
-
-	// AlignmentWithGoal measures how well the prompt serves its intended purpose
-	AlignmentWithGoal float64 `json:"alignmentWithGoal" validate:"required,min=0,max=20"`
+	OverallGrade      string       `json:"overall_grade" validate:"required,validGrade"`
+	Metrics           []Metric     `json:"metrics" validate:"required,min=1"`
+	Strengths         []Strength   `json:"strengths" validate:"required,min=1"`
+	Weaknesses        []Weakness   `json:"weaknesses" validate:"required,min=1"`
+	Suggestions       []Suggestion `json:"suggestions" validate:"required,min=1"`
+	OverallScore      float64      `json:"overall_score" validate:"required,min=0,max=20"`
+	EfficiencyScore   float64      `json:"efficiency_score" validate:"required,min=0,max=20"`
+	AlignmentWithGoal float64      `json:"alignment_with_goal" validate:"required,min=0,max=20"`
 }
 
 // OptimizationEntry represents a single step in the optimization process,
@@ -96,6 +69,8 @@ type OptimizationEntry struct {
 
 // OptimizerOption is a function type for configuring the PromptOptimizer.
 // It follows the functional options pattern for flexible configuration.
+//
+//nolint:revive // OptimizerOption follows the Option pattern naming convention commonly used in Go
 type OptimizerOption func(*PromptOptimizer)
 
 // IterationCallback is a function type for monitoring optimization progress.
@@ -106,45 +81,18 @@ type IterationCallback func(iteration int, entry OptimizationEntry)
 // It manages the iterative refinement of prompts through assessment,
 // improvement suggestions, and validation.
 type PromptOptimizer struct {
-	// llm is the language model used for optimization
-	llm llm.LLM
-
-	// debugManager handles debug output and tracking
-	debugManager *utils.DebugManager
-
-	// initialPrompt is the starting point for optimization
-	initialPrompt *llm.Prompt
-
-	// taskDesc describes the intended use of the prompt
-	taskDesc string
-
-	// customMetrics defines additional evaluation criteria
-	customMetrics []Metric
-
-	// optimizationGoal specifies the target outcome
-	optimizationGoal string
-
-	// history tracks the optimization process
-	history []OptimizationEntry
-
-	// ratingSystem defines the scoring methodology
-	ratingSystem string
-
-	// threshold sets the minimum acceptable score
-	threshold float64
-
-	// iterationCallback monitors optimization progress
+	llm               llm.LLM
 	iterationCallback IterationCallback
-
-	// maxRetries specifies retry attempts for failed operations
-	maxRetries int
-
-	// retryDelay sets the wait time between retries
-	retryDelay time.Duration
-
-	// memorySize limits the optimization history length
-	memorySize int
-
-	// iterations counts the optimization steps performed
-	iterations int
+	debugManager      *debug.Manager
+	initialPrompt     *llm.Prompt
+	taskDesc          string
+	optimizationGoal  string
+	ratingSystem      string
+	history           []OptimizationEntry
+	customMetrics     []Metric
+	threshold         float64
+	maxRetries        int
+	retryDelay        time.Duration
+	memorySize        int
+	iterations        int
 }
